feat(usb): add early handoff with PHY, SMMU, and trace support
Add modular USB gadget handoff support, including DWC3 controller
setup, PHY initialization, SMMU stream configuration, watchdog
handling, and a shared trace ring for early boot diagnostics.
Refactor inline UART logging into a dedicated module and expose
trace control helpers for host communication.
