+++ fullerene-kernel/src/arch/aarch64/usb.rs
@@ -8,7 +8,42 @@
 //! Early boot polls as a recovery path when firmware retains GIC ownership;
 //! the same event ring is drained from the IRQ handler once the GIC is live.
 
+use control::{
+    core_soft_reset, device_soft_reset, run_stop_device, stop_running_device, write_dctl_safe,
+};
 use core::ptr::{addr_of, addr_of_mut, read_volatile, write_volatile};
+
+use log::{log_hex, log_hex_value, log_puts};
+use mmio::*;
+use phy::{
+    init_hsphy, init_qmp_phy, qmp_set_autonomous_mode, select_utmi_pipe_clock,
+    update_dwc3_ref_clock,
+};
+mod control;
+mod phy;
+mod phy_tables;
+use trace::{fill_trace_control_response, trace_begin, trace_event};
+pub mod trace;
+use trace::*;
+pub use trace::{
+    TRACE_BOOT_USB_ENTRY, TRACE_EXCEPTION_SYNC, TRACE_PLATFORM_IRQ, TRACE_PROBE_WATCHDOG,
+    TRACE_TYPEC_BEGIN, TRACE_TYPEC_DONE, TRACE_TYPEC_EVENT, TRACE_UDC_REARM,
+    TRACE_USB_HANDOFF_BEGIN, dump_trace, trace_head, trace_last_event, trace_marker,
+    trace_probe_begin, trace_reset_head_for_boot,
+};
+mod log;
+mod mmio;
+mod smmu;
+mod watchdog;
+
+use smmu::{
+    adopt_smmu_dma_mapping, service_smmu_fault, smmu_install_stream_bypass, smmu_stream_s2cr_type,
+};
+pub use smmu::{configure_dwc3_smmu, probe_smmu_stream_state};
+use watchdog::{
+    CURRENT_EL_AT_ENTRY, MDCR_EL2_AT_ENTRY, SWDD_AVAIL, SWDD_RESULT, SWDD_STD, WDT_KPSS_EN_AT_ENTRY,
+};
+pub use watchdog::{secure_wdt_disable, secure_wdt_probes, u0_arm_wdt_bite, wdt_pet};
 
 use super::{
     uart,
@@ -20,878 +55,9 @@ use super::{
     usb_regs::*,
 };
 
-#[inline]
-fn log_puts(message: &str) {
-    #[cfg(not(fullerene_aarch64_usb_gadget_handoff_probe))]
-    uart::puts(message);
-}
-
-#[inline]
-fn log_hex(prefix: &str, value: u64) {
-    #[cfg(not(fullerene_aarch64_usb_gadget_handoff_probe))]
-    uart::put_hex(prefix, value);
-}
-
-#[inline]
-fn log_hex_value(value: u64) {
-    #[cfg(not(fullerene_aarch64_usb_gadget_handoff_probe))]
-    uart::put_hex_value(value);
-}
-
 unsafe extern "C" {
     static __usb_dma_start: u8;
     static __usb_dma_end: u8;
-}
-
-// Qualcomm APSS watchdog (kpss register layout). XBL/ABL arm it with a
-// pet-time of ~27 s; nothing in the early handoff pets it, so the handset
-// reboots (bark -> bite) ~8 s after the probe enters - exactly when host
-// enumeration is in flight. Pet it at entry and from every poll loop.
-const APSS_WDT_BASE: usize = 0x17c1_0000;
-// The SM7250 "qcom,wdt" (watchdog_v2) uses the kpss register layout at this
-// base (WDT0_RST=0x04, EN=0x08, STS=0x0C, BARK=0x10, BITE=0x14). The
-// apcs-timer offsets belong to other SoC families; writing them here would
-// hit reserved registers in the timer page.
-const APSS_WDT_RST: usize = 0x4;
-const APSS_WDT_EN: usize = 0x8;
-const APSS_WDT_STS: usize = 0xc;
-const APSS_WDT_BARK: usize = 0x10;
-const APSS_WDT_BITE: usize = 0x14;
-static mut WDT_TRACED: bool = false;
-// Non-zero while a scheduled APSS-WDT bite owns recovery: wdt_pet() must
-// stop petting (and must not re-arm the 100 s timeout) so the countdown
-// runs to the bite. The bite reboots the handset at a controlled time,
-// making the loop's return time a host-observable readout (the timing
-// readout the broken SMC software reset never delivered).
-static mut WDT_BITE_PENDING: u32 = 0;
-/// SCM result of the secure watchdog disable (0 = success; the sentinel
-/// means the call has not run yet or ran out of retries).
-static mut SWDD_RESULT: u64 = 0xFFFF_FFFF;
-/// SCM IS_CALL_AVAIL result for (SVC_BOOT, SEC_WDOG_DIS): 1 = the TZ
-/// implements the function, 0 = it answers but does not implement it,
-/// 0xFFFF_FFFF = it returned ARM_SMCCC_UNK_FUNC on every convention
-/// (the SMC path itself is dead, or the encoding is wrong). High word =
-/// convention attempt index (1 = SMC_64, 2 = SMC_32).
-static mut SWDD_AVAIL: u64 = 0xFFFF_FFFF;
-/// WDT enable word captured at the first pet (probe entry).
-/// 0xFFFF_FFFF = not captured yet.
-static mut WDT_KPSS_EN_AT_ENTRY: u32 = 0xFFFF_FFFF;
-
-#[inline]
-unsafe fn wdt_reg(offset: usize) -> *mut u32 {
-    (APSS_WDT_BASE + offset) as *mut u32
-}
-
-/// Deactivate the SECURE watchdog through SCM (the same call the downstream
-/// watchdog_v2 driver exposes via its disable sysfs).
-///
-/// The bramble-generation TZ (msm-5.4 class) speaks the ARM SMCCC vendor
-/// interface (driver `qcom_scm-64.c` + `watchdog_v2.c`):
-///   x0 = ARM_SMCCC_CALL_VAL(ARM_SMCCC_STD_CALL, <convention>,
-///                           ARM_SMCCC_OWNER_SIP,
-///                           QCOM_SCM_FNID(QCOM_SCM_SVC_BOOT,
-///                                         QCOM_SCM_BOOT_SEC_WDOG_DIS))
-///      = 0x00 | {0x20 (SMC_64) | 0x00 (SMC_32)} | 0x40 | ((0x1 << 8) | 0x7)
-///   x1 = arginfo = 1, x2 = args[0] = 1, x3..x7 = 0.
-/// The vendor driver PROBES the TZ's convention at runtime, so try SMC_64
-/// (0x0167) first and fall back to SMC_32 (0x0147). XBL/ABL arm this
-/// watchdog for the `fastboot boot` path; an unpetted bite reboots the
-/// handset ~17 s into every probe (bootreason=watchdog) no matter what
-/// happens to the APSS WDT registers. If TZ is servicing another client it
-/// answers QCOM_SCM_INTERRUPTED (1) and expects the call to be reissued
-/// with x0 = 1 and the x6 continuation value it returned. The trace/`SWDD`
-/// record keeps the attempt index (1 = SMC_64, 2 = SMC_32) in the upper
-/// word of `SWDD_RESULT`.
-/// SMCCC_VERSION (ARM_SMCCC_VERSION = 0x80000000) result captured at
-/// probe entry: a nonzero (major<<16|minor) value proves the EL3 SMCCC
-/// dispatch answers at all; 0xFFFF_FFFF means every SMC is faked or
-/// trapped (the answer never came from a SMCCC handler).
-static mut SWDD_STD: u64 = 0xFFFF_FFFF;
-/// MDCR_EL2 (Memory Debug Control, EL2) captured at probe entry. Bit 14
-/// (SMC) traps SMC instructions issued from EL0/EL1 into EL2: if the
-/// bootloader left it set, every "SMC" we issue lands in the bootloader's
-/// EL2 vector, never in TZ, and whatever that vector does (often a
-/// default -1 return) is all we ever see.
-static mut MDCR_EL2_AT_ENTRY: u64 = u64::MAX;
-/// CurrentEL at probe entry: 0b0101 = EL1h (expected), 0b1000 = EL2h.
-static mut CURRENT_EL_AT_ENTRY: u64 = u64::MAX;
-
-/// One QCOM SCM SMC call with the QCOM_SCM_INTERRUPTED retry loop: on a
-/// return of 1 the call is reissued with x0 = 1 and x6 = the continuation
-/// value the firmware stored in a6 (the ARM_SMCCC_QUIRK_QCOM_A6 pattern).
-/// Returns (a0, a6) of the final answer.
-unsafe fn scm_smc_call_once(fnid: u64, arginfo: u64, a0: u64, a1: u64, a2: u64) -> (u64, u64) {
-    let mut fnid = fnid;
-    let mut a6: u64 = 0;
-    let mut result: u64 = 0xFFFF_FFFF;
-    for _ in 0..100 {
-        let mut a6_out: u64 = 0;
-        core::arch::asm!(
-            "mov x0, {fnid}",
-            "mov x1, {arginfo}",
-            "mov x2, {a0v}",
-            "mov x3, {a1v}",
-            "mov x4, {a2v}",
-            "mov x5, xzr",
-            "mov x6, {a6in}",
-            "mov x7, xzr",
-            "smc #0",
-            "mov {result}, x0",
-            "mov {a6out}, x6",
-            fnid = in(reg) fnid,
-            arginfo = in(reg) arginfo,
-            a0v = in(reg) a0,
-            a1v = in(reg) a1,
-            a2v = in(reg) a2,
-            a6in = in(reg) a6,
-            result = out(reg) result,
-            a6out = out(reg) a6_out,
-            out("x1") _,
-            out("x2") _,
-            out("x3") _,
-            out("x4") _,
-            out("x5") _,
-            out("x7") _,
-            options(nostack)
-        );
-        if result != 1 {
-            break;
-        }
-        // QCOM_SCM_INTERRUPTED: reissue with the continuation value.
-        fnid = 1;
-        a6 = a6_out;
-    }
-    (result, a6)
-}
-
-/// ARM_SMCCC_CALL_VAL(ARM_SMCCC_STD_CALL, conv, OWNER_SIP=2, fn) per the
-/// 5.4 qcom_scm driver: TYPE(31)=0, CONV(30), OWNER(24..29)=2, FN(0..15).
-#[inline]
-fn scm_oen(conv: u64, fnid: u32) -> u64 {
-    (conv << 30) | (2u64 << 24) | (fnid as u64)
-}
-
-pub fn secure_wdt_disable() -> bool {
-    // One legacy-encoding QCOM SMC at probe entry: SMC id = 0x02000000 |
-    // (SCM_SVC_BOOT << 10) | SCM_SVC_SEC_WDOG_DIS = 0x02000407, arginfo = 1,
-    // args[0] = 1. This is the minimal entry that the fastboot handoff has
-    // been proven to tolerate on this target: the extended multi-SMC
-    // diagnostic sequence issued at entry (SMCCC VERSION, IS_CALL_AVAIL,
-    // both-convention SEC_WDOG_DIS) deterministically wedges the handoff in
-    // the standalone probe, so those probes live in secure_wdt_probes() and
-    // run only in the signal-probe build, where the gate consumes them.
-    // The fnid is build-selectable so the A/B run can swap encodings without
-    // a source change: the default is the proven-to-be-tolerated legacy call
-    // (0x02000407); the modern SMCCC OEN encodings of the same function
-    // (BOOT/0x07, funcnum 0x0107, owner SIP) are selected by
-    // FULLERENE_USB_SWDD_FNID, e.g. 0x82000107 = STD/SMC64, 0x02000107 =
-    // FAST/SMC64, 0xA2000107 = STD/SMC32, 0x22000107 = FAST/SMC32. The upper
-    // word of SWDD_RESULT records the fnid actually issued (the lower word
-    // is the result, so the swdd-ok gate keeps its meaning).
-    let fnid = option_env!("FULLERENE_USB_SWDD_FNID")
-        .and_then(|value| u64::from_str_radix(value.trim_start_matches("0x"), 16).ok())
-        .unwrap_or(0x0200_0407);
-    let result = unsafe { scm_smc_call_once(fnid, 1, 1, 0, 0) }.0;
-    unsafe {
-        SWDD_RESULT = (fnid << 32) | (result & 0xFFFF_FFFF);
-    }
-    trace_event(
-        TRACE_PROBE_WATCHDOG,
-        0x5357_4444, // "SWDD"
-        result as u32,
-        (fnid >> 16) as u32,
-        fnid as u32,
-        0,
-    );
-    result == 0
-}
-
-/// Extended SMCCC diagnostics consumed by the `swdd-*` / `scm-*` / `std-*` /
-/// `mdcr-*` / `el1` / `el2` signal gates. Issued at probe entry, BEFORE any
-/// SMC that the standalone build performs at entry is multiplied out, and
-/// never in the standalone build (its multi-SMC entry sequence wedges the
-/// fastboot handoff; see secure_wdt_disable).
-pub fn secure_wdt_probes() {
-    // Capture the exception-level context BEFORE issuing any SMC: if the
-    // bootloader left MDCR_EL2.SMC (bit 14) set, SMC from EL1 traps to EL2
-    // and never reaches TZ, which would fake every one of these probes.
-    unsafe {
-        let mdcr: u64;
-        let cur_el: u64;
-        core::arch::asm!(
-            "mrs {m}, MDCR_EL2",
-            "mrs {c}, CurrentEL",
-            m = out(reg) mdcr,
-            c = out(reg) cur_el,
-            options(nostack)
-        );
-        MDCR_EL2_AT_ENTRY = mdcr;
-        CURRENT_EL_AT_ENTRY = cur_el;
-        trace_event(
-            TRACE_PROBE_WATCHDOG,
-            0x4D44_4352, // "MDCR"
-            mdcr as u32,
-            (mdcr >> 32) as u32,
-            cur_el as u32,
-            0,
-        );
-    }
-
-    // Baseline probe: ARM SMCCC VERSION (OEN 0x80000000, no args). Per the
-    // SMCCC spec every compliant EL3 handler must answer this, so it
-    // separates "the SMC never reaches a SMCCC handler" from "the handler
-    // works but does not carry the QCOM SIP service table".
-    let std_result = unsafe { scm_smc_call_once(0x8000_0000, 0, 0, 0, 0) }.0;
-    unsafe {
-        SWDD_STD = std_result;
-    }
-    trace_event(
-        TRACE_PROBE_WATCHDOG,
-        0x5343_4D53, // "SCMS"
-        std_result as u32,
-        0,
-        0,
-        0,
-    );
-
-    // Diagnostic probe: QCOM_SCM_IS_CALL_AVAIL(SVC_INFO=0x06, CMD=0x01) with
-    // args (SVC_BOOT, SEC_WDOG_DIS). The vendor driver itself calls this to
-    // probe the convention, so the function is guaranteed present in a
-    // working TZ. Its answer separates "SMC path dead / encoding wrong"
-    // (UNK_FUNC = -1) from "the SMC works but the SEC_WDOG_DIS function is
-    // absent" (0). fnid = (0x06<<8)|0x01 = 0x0601; args = (SVC_BOOT, 0x07).
-    let mut avail_attempt: u32 = 0;
-    let mut avail: u64 = 0xFFFF_FFFF;
-    for conv in [scm_oen(1, 0x0601), scm_oen(0, 0x0601)] {
-        avail_attempt += 1;
-        avail = unsafe { scm_smc_call_once(conv, 2, 0x01, 0x07, 0) }.0;
-        if avail != 0xFFFF_FFFF {
-            break;
-        }
-    }
-    unsafe {
-        SWDD_AVAIL = ((avail_attempt as u64) << 32) | avail;
-    }
-    trace_event(
-        TRACE_PROBE_WATCHDOG,
-        0x5343_4D41, // "SCMA"
-        avail as u32,
-        avail_attempt,
-        0,
-        0,
-    );
-}
-
-/// Restart the apps watchdog countdown and, once, record its configuration
-/// in the retained trace, disabling it if it is armed.
-pub fn wdt_pet() {
-    unsafe {
-        if WDT_BITE_PENDING != 0 {
-            // A scheduled bite owns recovery now: any pet (including the
-            // one-shot 100 s re-arm below) would cancel it.
-            return;
-        }
-        if !WDT_TRACED {
-            WDT_TRACED = true;
-            let en = read_volatile(wdt_reg(APSS_WDT_EN));
-            WDT_KPSS_EN_AT_ENTRY = en;
-            let bark = read_volatile(wdt_reg(APSS_WDT_BARK));
-            let bite = read_volatile(wdt_reg(APSS_WDT_BITE));
-            let sts = read_volatile(wdt_reg(APSS_WDT_STS));
-            trace_event(
-                TRACE_PROBE_WATCHDOG,
-                0x5744_5430, // "WDT0"
-                en,
-                bark,
-                bite,
-                sts,
-            );
-            // Re-arm the countdown with OUR timeout: the downstream
-            // watchdog_v2 configures bark = qcom,bark-time (+console offset)
-            // and bite = bark + 3 s at a 32765 Hz watchdog clock. Writing a
-            // 100 s bark / 110 s bite here moves an APSS-WDT-driven reboot
-            // out of the enumeration window entirely; if the ~17 s reboot
-            // still happens, the source is a different watchdog.
-            let bark_100s: u32 = 100 * 32765;
-            let bite_110s: u32 = 110 * 32765;
-            write_volatile(wdt_reg(APSS_WDT_BARK), bark_100s);
-            write_volatile(wdt_reg(APSS_WDT_BITE), bite_110s);
-            write_volatile(wdt_reg(APSS_WDT_RST), 1);
-            write_volatile(wdt_reg(APSS_WDT_EN), 1);
-            core::arch::asm!("dsb sy", options(nostack));
-            let en_after = read_volatile(wdt_reg(APSS_WDT_EN));
-            trace_event(
-                TRACE_PROBE_WATCHDOG,
-                0x5744_5441, // "WDTA" armed by probe
-                en,
-                en_after,
-                bark_100s >> 16,
-                bite_110s >> 16,
-            );
-        }
-        // The downstream watchdog_v2 pets by writing 1 (not 0) to RST.
-        write_volatile(wdt_reg(APSS_WDT_RST), 1);
-        core::arch::asm!("dsb sy", options(nostack));
-    }
-}
-
-/// Schedule the APSS watchdog bite `delay` seconds from now: re-arm
-/// bark/bite at the 32765 Hz watchdog clock, restart the countdown, and
-/// mark the pet path so nothing cancels it. The bite reboots the handset
-/// on its own (no SMC software reset needed), so the loop's return time
-/// — bite time plus the ~20 s Android boot — becomes a host-observable
-/// readout of which u0_arm_recovery step failed. If the APSS WDT is not
-/// writable from here (secure-owned), the bite never lands and the secure
-/// watchdog's ~17 s bite (the ~37 s return) stands in for every bucket.
-/// The callers own the gating (the u0-arm status match is unreachable
-/// without the u0-arm env; the control bite has its own env).
-pub fn u0_arm_wdt_bite(delay_secs: u32) {
-    unsafe {
-        let delay = delay_secs.clamp(1, 16);
-        let bark = (delay.saturating_sub(1).max(1) * 32765) as u32;
-        let bite = (delay * 32765) as u32;
-        write_volatile(wdt_reg(APSS_WDT_BARK), bark);
-        write_volatile(wdt_reg(APSS_WDT_BITE), bite);
-        write_volatile(wdt_reg(APSS_WDT_RST), 1);
-        write_volatile(wdt_reg(APSS_WDT_EN), 1);
-        core::arch::asm!("dsb sy", options(nostack));
-        WDT_BITE_PENDING = delay;
-        trace_marker(TRACE_PROBE_WATCHDOG, 0x5744_5442 | (delay & 0xff)); // "WDTB"
-    }
-}
-
-#[inline]
-fn dwc3_base() -> usize {
-    super::platform::bramble::usb_resources().dwc3_base
-}
-// Lito/SM7250's Apps SMMU owns the DWC3 stream ID declared by the board DT.
-// The early Bramble path installs a small identity map in a context bank so
-// the USB buffers remain inside the IOVA pool declared by the vendor DT.
-// Google’s Bramble/Lito DTS places apps-smmu at 0x15000000.  The nearby
-// 0x0c600000 range is the SPMI arbiter channel window, not the Apps SMMU;
-// confusing the two makes the SMMU identity-map setup target unrelated PMIC
-// registers before the first EP0 transfer.
-const SMMU_ID0: usize = 0x20;
-const SMMU_ID1: usize = 0x24;
-const SMMU_ID0_NUMSMRG_MASK: u32 = 0xff;
-const SMMU_ID1_PAGESIZE: u32 = 1 << 31;
-const SMMU_ID1_NUMPAGENDXB_SHIFT: u32 = 28;
-const SMMU_ID1_NUMPAGENDXB_MASK: u32 = 0x7;
-const SMMU_ID1_NUMS2CB_SHIFT: u32 = 16;
-const SMMU_ID1_NUMS2CB_MASK: u32 = 0xff;
-const SMMU_ID1_NUMCB_MASK: u32 = 0xff;
-const SMMU_SMR_BASE: usize = 0x800;
-const SMMU_S2CR_BASE: usize = 0xc00;
-const SMMU_TLB_ALL_H: usize = 0x6c;
-const SMMU_TLB_SYNC: usize = 0x70;
-const SMMU_TLB_STATUS: usize = 0x74;
-const SMMU_TLB_STATUS_ACTIVE: u32 = 1;
-const SMMU_SMR_VALID: u32 = 1 << 31;
-const SMMU_SMR_MASK_SHIFT: u32 = 16;
-const SMMU_S2CR_TYPE_MASK: u32 = 0x3 << 16;
-const SMMU_S2CR_TYPE_TRANS: u32 = 0;
-const SMMU_S2CR_TYPE_BYPASS: u32 = 1 << 16;
-const SMMU_S2CR_CBNDX_MASK: u32 = 0xff;
-const SMMU_GR1_CBAR_BASE: usize = 0x00;
-const SMMU_GR1_CBA2R_BASE: usize = 0x800;
-const SMMU_CBA2R_VA64: u32 = 1;
-const SMMU_CBAR_IRPTNDX_MASK: u32 = 0xff;
-const SMMU_CBAR_S1_TRANS_S2_BYPASS: u32 = 1 << 16;
-const SMMU_CBAR_S1_MEMATTR_WB: u32 = 0xf << 12;
-const SMMU_CBAR_S1_BPSHCFG_NSH: u32 = 3 << 8;
-const SMMU_CB_SCTLR: usize = 0x00;
-const SMMU_CB_TCR2: usize = 0x10;
-const SMMU_CB_TTBR0: usize = 0x20;
-const SMMU_CB_TTBR1: usize = 0x28;
-const SMMU_CB_TCR: usize = 0x30;
-const SMMU_CB_CONTEXTIDR: usize = 0x34;
-const SMMU_CB_MAIR0: usize = 0x38;
-const SMMU_CB_MAIR1: usize = 0x3c;
-const SMMU_CB_RESUME: usize = 0x08;
-const SMMU_CB_FSR: usize = 0x58;
-const SMMU_CB_FAR: usize = 0x60;
-const SMMU_CB_FSYNR0: usize = 0x68;
-const SMMU_GR0_FSR: usize = 0x48;
-const SMMU_GR0_FSYNR0: usize = 0x50;
-const SMMU_RESUME_TERMINATE: u32 = 1;
-const SMMU_GLOBAL_FSR_FAULT: u32 = 1 << 1;
-const SMMU_FSR_SS: u32 = 1 << 30;
-const SMMU_FSR_FAULT: u32 = (1 << 31)
-    | (1 << 30)
-    | (1 << 8)
-    | (1 << 7)
-    | (1 << 6)
-    | (1 << 5)
-    | (1 << 4)
-    | (1 << 3)
-    | (1 << 2)
-    | (1 << 1);
-const SMMU_SCTLR_S1_ASIDPNE: u32 = 1 << 12;
-const SMMU_SCTLR_CFIE: u32 = 1 << 6;
-const SMMU_SCTLR_CFRE: u32 = 1 << 5;
-const SMMU_SCTLR_AFE: u32 = 1 << 2;
-const SMMU_SCTLR_TRE: u32 = 1 << 1;
-const SMMU_SCTLR_M: u32 = 1;
-const SMMU_GR0_SCR0: usize = 0x00;
-const SMMU_SCR0_GFRE: u32 = 1 << 1;
-const SMMU_SCR0_GFIE: u32 = 1 << 2;
-const SMMU_SCR0_GCFGFRE: u32 = 1 << 4;
-const SMMU_SCR0_GCFGFIE: u32 = 1 << 5;
-const SMMU_TCR_EPD1: u32 = 1 << 23;
-const SMMU_TCR_SH0_INNER: u32 = 3 << 12;
-const SMMU_TCR_ORGN0_WBWA: u32 = 1 << 10;
-const SMMU_TCR_IRGN0_WBWA: u32 = 1 << 8;
-const SMMU_TCR_T0SZ_32BIT: u32 = 32;
-const SMMU_TCR_T0SZ_39BIT: u32 = 25;
-const SMMU_TCR2_SEP_UPSTREAM: u32 = 0x7 << 15;
-const SMMU_TCR2_AS: u32 = 1 << 4;
-const SMMU_TCR2_PASIZE_40BIT: u32 = 2;
-
-#[repr(C, align(4096))]
-#[derive(Clone, Copy)]
-struct SmmuTable([u64; 512]);
-
-// With T0SZ=32 and a 4 KiB granule, TTBR0 points at a level-1 table. The
-// vendor DT's 0x90000000..0xf0000000 pool is represented by level-2 2 MiB
-// identity blocks; addresses outside the pool are left unmapped. These tables
-// are cleared together with the other USB DMA objects.
-#[unsafe(link_section = ".usb_dma")]
-static mut SMMU_L1: SmmuTable = SmmuTable([0; 512]);
-#[unsafe(link_section = ".usb_dma")]
-static mut SMMU_L2: [SmmuTable; 4] = [SmmuTable([0; 512]); 4];
-static mut SMMU_CONTEXT_PAGE: usize = usize::MAX;
-static mut SMMU_CONTEXT_PAGE_SIZE: usize = 0;
-
-const SMMU_DESC_VALID: u64 = 1;
-const SMMU_DESC_TABLE: u64 = 3;
-const SMMU_DESC_BLOCK: u64 = 1;
-const SMMU_DESC_TYPE_MASK: u64 = 3;
-const SMMU_DESC_ADDRESS_MASK: u64 = 0x000f_ffff_ffff_f000;
-const SMMU_DESC_AF: u64 = 1 << 10;
-const SMMU_DESC_SH_INNER: u64 = 3 << 8;
-const SMMU_DESC_ATTR_NORMAL: u64 = 0;
-const SMMU_DESC_XN: u64 = (1 << 53) | (1 << 54);
-#[inline]
-fn apps_smmu_base() -> usize {
-    super::platform::bramble::usb_resources().apps_smmu_base
-}
-
-#[inline]
-fn hsphy_base() -> usize {
-    super::platform::bramble::usb_resources().hs_phy_base
-}
-
-#[inline]
-fn qmp_base() -> usize {
-    super::platform::bramble::usb_resources().qmp_phy_base
-}
-// SM7250 exposes the Qualcomm glue/QSCRATCH block immediately above the
-// DWC3 core.  The glue must report the cable's VBUS/session to the core when
-// we take over directly from the bootloader.
-#[inline]
-fn qscratch_base() -> usize {
-    super::platform::bramble::usb_resources().qscratch_base
-}
-const QSCRATCH_HS_PHY_CTRL: usize = 0x10;
-const QSCRATCH_CGCTL: usize = 0x28;
-const QSCRATCH_SS_PHY_CTRL: usize = 0x30;
-const QSCRATCH_GENERAL_CFG: usize = 0x08;
-const QSCRATCH_GENERAL_CFG_XHCI_REV: u32 = 1 << 2;
-// Qualcomm glue power-event status/mask registers. These are consumed by
-// dwc3-msm's threaded power IRQ, not by the DWC3 event ring.
-const QSCRATCH_PWR_EVENT_STATUS: usize = 0x58;
-const QSCRATCH_PWR_EVENT_MASK: usize = 0x5c;
-const PWR_EVENT_POWERDOWN_IN_P3: u32 = 1 << 2;
-const PWR_EVENT_POWERDOWN_OUT_P3: u32 = 1 << 3;
-const PWR_EVENT_LPM_IN_L2: u32 = 1 << 4;
-const PWR_EVENT_LPM_OUT_L2: u32 = 1 << 5;
-const PWR_EVENT_LPM_OUT_L1: u32 = 1 << 13;
-
-const GCTL: usize = 0xc110;
-const GUCTL: usize = 0xc12c;
-const GUCTL2: usize = 0xc19c;
-// DWC3_GUCTL1 is part of the global register block immediately after GCTL;
-// 0xc360 is in the FIFO-register area and is not a user-control register.
-const GUCTL1: usize = 0xc11c;
-// msm-4.19 core.h: GUCTL3 sits in the 0xc600 block, not next to
-// GUCTL1; the 4.19 offset is 0xc60c.
-const GUCTL3: usize = 0xc60c;
-const GSNPSID: usize = 0xc120;
-const GRXTHRCFG: usize = 0xc10c;
-const GSBUSCFG1: usize = 0xc104;
-const GHWPARAMS0: usize = 0xc140;
-const GHWPARAMS1: usize = 0xc144;
-const GHWPARAMS3: usize = 0xc14c;
-const GHWPARAMS7: usize = 0xc15c;
-const GDBGLTSSM: usize = 0xc164;
-const VER_NUMBER: usize = 0xc1a0;
-const VER_TYPE: usize = 0xc1a4;
-const GFLADJ: usize = 0xc630;
-const GUSB2PHYCFG0: usize = 0xc200;
-const GUSB3PIPECTL0: usize = 0xc2c0;
-const GUSB2PHYCFG_ULPI_UTMI: u32 = 1 << 4;
-const GUSB2PHYCFG_PHYIF_MASK: u32 = 1 << 3;
-const GUSB2PHYCFG_USBTRDTIM_MASK: u32 = 0xf << 10;
-const GUSB2PHYCFG_USBTRDTIM_UTMI_8_BIT: u32 = 9 << 10;
-const GEVNTADRLO0: usize = 0xc400;
-const GEVNTADRHI0: usize = 0xc404;
-const GEVNTSIZ0: usize = 0xc408;
-const GEVNTCOUNT0: usize = 0xc40c;
-const DEV_IMOD0: usize = 0xca00;
-const GEVNT_BUFFER_STRIDE: usize = 0x10;
-const DCFG: usize = 0xc700;
-const DCTL: usize = 0xc704;
-const DEVTEN: usize = 0xc708;
-const DSTS: usize = 0xc70c;
-const DALEPENA: usize = 0xc720;
-const DEP_BASE: usize = 0xc800;
-
-const GCTL_PRTCAPDIR_MASK: u32 = 3 << 12;
-const GCTL_PRTCAP_DEVICE: u32 = 2 << 12;
-const GCTL_U2RSTECN: u32 = 1 << 16;
-const GCTL_SCALEDOWN_MASK: u32 = 3 << 4;
-const GCTL_DISSCRAMBLE: u32 = 1 << 3;
-const GCTL_CORESOFTRESET: u32 = 1 << 11;
-const GCTL_GBLHIBERNATIONEN: u32 = 1 << 1;
-const GHWPARAMS1_EN_PWROPT_MASK: u32 = 3 << 24;
-const GHWPARAMS1_EN_PWROPT_HIB: u32 = 2 << 24;
-const GCTL_DSBLCLKGTNG: u32 = 1;
-const GUCTL_REFCLKPER_MASK: u32 = 0xffc0_0000;
-const GUCTL_REFCLKPER_19_2MHZ: u32 = 52 << 22;
-const GFLADJ_REFCLK_FLADJ_MASK: u32 = 0x003f_ff00;
-const GFLADJ_REFCLK_LPM_SEL: u32 = 1 << 23;
-const GFLADJ_REFCLK_240MHZ_DECR: u32 = 12 << 24;
-const GFLADJ_REFCLK_240MHZDECR_PLS1: u32 = 1 << 31;
-const GFLADJ_REFCLK_FLADJ_19_2MHZ: u32 = 200 << 8;
-const GUSB2PHYCFG_SUSPHY: u32 = 1 << 6;
-const GUSB2PHYCFG_ENBLSLPM: u32 = 1 << 8;
-const GUSB2PHYCFG_PHYSOFTRST: u32 = 1 << 31;
-const GUCTL1_L1_SUSP_THRLD_EN_FOR_HOST: u32 = 1 << 8;
-const GUCTL1_DEV_L1_EXIT_BY_HW: u32 = 1 << 24;
-const GUCTL1_IP_GAP_ADD_ON: u32 = 1 << 21;
-const GUCTL3_USB20_RETRY_DISABLE: u32 = 1 << 16;
-const GSBUSCFG1_PIPETRANSLIMIT_MASK: u32 = 0x0f << 8;
-const GSBUSCFG1_PIPETRANSLIMIT_E: u32 = 0xe << 8;
-const GUSB3PIPECTL_SUSPHY: u32 = 1 << 17;
-const GUSB3PIPECTL_PHYSOFTRST: u32 = 1 << 31;
-
-const DCTL_CSFTRST: u32 = 1 << 30;
-const DCTL_SDIS: u32 = 1 << 0;
-const DCTL_HIRD_THRES_MASK: u32 = 0x1f << 24;
-const DCTL_HIRD_THRES_LITO: u32 = 0x10 << 24;
-const DCTL_KEEP_CONNECT: u32 = 1 << 19;
-const DCTL_TRGTULST_MASK: u32 = 0x0f << 17;
-const DCTL_TRGTULST_RX_DET: u32 = 5 << 17;
-const DCFG_NUMP_SHIFT: u32 = 17;
-const DCFG_NUMP_MASK: u32 = 0x1f << DCFG_NUMP_SHIFT;
-const DCFG_IGNSTRMPP: u32 = 1 << 23;
-const DWC3_GRXTHRCFG_PKTCNTSEL: u32 = 1 << 29;
-const DWC31_GRXTHRCFG_PKTCNTSEL: u32 = 1 << 26;
-const GHWPARAMS0_MDWIDTH_SHIFT: u32 = 8;
-const GHWPARAMS0_MDWIDTH_MASK: u32 = 0xff;
-const GHWPARAMS7_RAM2_DEPTH_SHIFT: u32 = 16;
-const GHWPARAMS7_RAM2_DEPTH_MASK: u32 = 0xffff;
-const DWC3_IP: u32 = 0x5533;
-const DWC31_IP: u32 = 0x3331;
-const DWC32_IP: u32 = 0x3332;
-const DWC31_REVISION_180A: u32 = 0x3138_302a;
-const DWC31_REVISION_190A: u32 = 0x3139_302a;
-// Linux applies the RxDetect reconnect workaround only through DWC3 1.87a.
-// GSNPSID carries the same full revision value used by the upstream driver.
-const DWC3_REVISION_187A: u32 = 0x5533_187a;
-const DWC3_REVISION_190A: u32 = 0x5533_190a;
-const DWC3_REVISION_194A: u32 = 0x5533_194a;
-const DWC3_REVISION_220A: u32 = 0x5533_220a;
-const DWC3_REVISION_250A: u32 = 0x5533_250a;
-const DWC3_REVISION_310A: u32 = 0x5533_310a;
-const DWC3_REVISION_270A: u32 = 0x5533_270a;
-const DWC3_REVISION_290A: u32 = 0x5533_290a;
-// msm-4.19 core.c dwc3_core_is_valid(): a DWC_usb31 core reports its
-// revision as VER_NUMBER with the high bit set (1.70a = 0x3137302a).
-const DWC3_REVISION_IS_DWC31: u32 = 0x8000_0000;
-const DWC3_USB31_REVISION_170A: u32 = 0x3137_302a | DWC3_REVISION_IS_DWC31;
-// VER_TYPE "ga**" marks the general-availability usb31 silicon.
-const DWC3_USB31_VER_TYPE_GA: u32 = 0x6761_2a2a;
-const GUCTL2_RST_ACTBITLATER: u32 = 1 << 14;
-
-const HSPHY_UTMI_CTRL0: usize = 0x3c;
-const HSPHY_UTMI_CTRL5: usize = 0x50;
-const HSPHY_COMMON0: usize = 0x54;
-const HSPHY_COMMON1: usize = 0x58;
-const HSPHY_COMMON2: usize = 0x5c;
-const HSPHY_CTRL1: usize = 0x60;
-const HSPHY_CTRL2: usize = 0x64;
-const HSPHY_CFG0: usize = 0x94;
-const HSPHY_REFCLK_CTRL: usize = 0xa0;
-const HSPHY_RTUNE_SEL: usize = 0xb4;
-const HSPHY_TEST0: usize = 0x80;
-const HSPHY_TEST1: usize = 0x84;
-
-const HSPHY_UTMI_SLEEPM: u32 = 1 << 0;
-const HSPHY_UTMI_ATE_RESET: u32 = 1 << 0;
-const HSPHY_UTMI_POR: u32 = 1 << 1;
-const HSPHY_COMMON0_FSEL_MASK: u32 = 0x7 << 4;
-const HSPHY_COMMON0_VATESTENB_MASK: u32 = 0x3;
-const HSPHY_COMMON1_VBUSVLDEXTSEL0: u32 = 1 << 4;
-const HSPHY_COMMON1_PLLBTUNE: u32 = 1 << 5;
-const HSPHY_COMMON2_VREGBYPASS: u32 = 1 << 0;
-const HSPHY_CTRL1_VBUSVLDEXT0: u32 = 1 << 0;
-const HSPHY_CTRL2_SUSPEND_N: u32 = 1 << 2;
-const HSPHY_CTRL2_SUSPEND_N_SEL: u32 = 1 << 3;
-const HSPHY_CFG0_CMN_CTRL_OVERRIDE_EN: u32 = 1 << 1;
-const HSPHY_TEST1_TESTDATAOUTSEL: u32 = 1 << 4;
-const HSPHY_TEST1_TOGGLE_2WR: u32 = 1 << 6;
-const HSPHY_TEST0_DATA_MASK: u32 = 0xff;
-
-const PIPE_UTMI_CLK_SEL: u32 = 1 << 0;
-const PIPE3_PHYSTATUS_SW: u32 = 1 << 3;
-const PIPE_UTMI_CLK_DIS: u32 = 1 << 8;
-
-// Qualcomm's Android wrapper reserves event buffers 1..N for GSI. These
-// fields are part of the DWC3 event-buffer ABI, not ordinary endpoint
-// registers, so keep the encoding next to the event-ring setup.
-const GSI_TRB_ADDR_BIT_53: u32 = 1 << 21;
-const GSI_TRB_ADDR_BIT_55: u32 = 1 << 23;
-const GSI_CLK_EN: u32 = 1 << 12;
-const GSI_RESTART_DBL_PNTR: u32 = 1 << 20;
-const GSI_EN: u32 = 1 << 0;
-const GSI_BLOCK_WR_GO: u32 = 1 << 1;
-const GSI_EVENT_INTR_MASK: u32 = 1 << 31;
-const GSI_EVENT_ADDR_EN_SHIFT: u32 = 22;
-const GSI_EVENT_ADDR_INDEX_SHIFT: u32 = 16;
-const GSI_WR_CTRL_STATE: u32 = 1 << 15;
-
-const QMP_COM_PHY_MODE_CTRL: usize = 0x0000;
-const QMP_COM_SW_RESET: usize = 0x0004;
-const QMP_COM_POWER_DOWN_CTRL: usize = 0x0008;
-const QMP_COM_TYPEC_CTRL: usize = 0x0010;
-const QMP_COM_RESET_OVRD_CTRL: usize = 0x001c;
-const QMP_PCS_STATUS1: usize = 0x1c14;
-const QMP_PCS_AUTONOMOUS_MODE_CTRL: usize = 0x1f08;
-const QMP_PCS_LFPS_RXTERM_IRQ_CLEAR: usize = 0x1f14;
-const QMP_PCS_CLAMP_ENABLE: usize = 0x1c8c;
-const QMP_PCS_POWER_DOWN_CONTROL: usize = 0x1c40;
-const QMP_PCS_SW_RESET: usize = 0x1c00;
-const QMP_PCS_START_CONTROL: usize = 0x1c44;
-const QMP_PHYSTATUS: u32 = 1 << 6;
-const QMP_ARCVR_DTCT_EN: u32 = 1 << 0;
-const QMP_ALFPS_DTCT_EN: u32 = 1 << 1;
-const QMP_ARCVR_DTCT_EVENT_SEL: u32 = 1 << 4;
-const QMP_LFPS_IRQ_CLEAR: u32 = 1 << 0;
-const QMP_CLAMP_EN: u32 = 1 << 0;
-
-const QMP_INIT: [(usize, u32); 146] = [
-    (0x1010, 0x01), // USB3_DP_QSERDES_COM_SSC_EN_CENTER
-    (0x101c, 0x31), // USB3_DP_QSERDES_COM_SSC_PER1
-    (0x1020, 0x01), // USB3_DP_QSERDES_COM_SSC_PER2
-    (0x1024, 0xde), // USB3_DP_QSERDES_COM_SSC_STEP_SIZE1_MODE0
-    (0x1028, 0x07), // USB3_DP_QSERDES_COM_SSC_STEP_SIZE2_MODE0
-    (0x1030, 0xde), // USB3_DP_QSERDES_COM_SSC_STEP_SIZE1_MODE1
-    (0x1034, 0x07), // USB3_DP_QSERDES_COM_SSC_STEP_SIZE2_MODE1
-    (0x1050, 0x0a), // USB3_DP_QSERDES_COM_SYSCLK_BUF_ENABLE
-    (0x1060, 0x20), // USB3_DP_QSERDES_COM_CMN_IPTRIM
-    (0x1074, 0x06), // USB3_DP_QSERDES_COM_CP_CTRL_MODE0
-    (0x1078, 0x06), // USB3_DP_QSERDES_COM_CP_CTRL_MODE1
-    (0x107c, 0x16), // USB3_DP_QSERDES_COM_PLL_RCTRL_MODE0
-    (0x1080, 0x16), // USB3_DP_QSERDES_COM_PLL_RCTRL_MODE1
-    (0x1084, 0x36), // USB3_DP_QSERDES_COM_PLL_CCTRL_MODE0
-    (0x1088, 0x36), // USB3_DP_QSERDES_COM_PLL_CCTRL_MODE1
-    (0x1094, 0x1a), // USB3_DP_QSERDES_COM_SYSCLK_EN_SEL
-    (0x10a4, 0x04), // USB3_DP_QSERDES_COM_LOCK_CMP_EN
-    (0x10ac, 0x14), // USB3_DP_QSERDES_COM_LOCK_CMP1_MODE0
-    (0x10b0, 0x34), // USB3_DP_QSERDES_COM_LOCK_CMP2_MODE0
-    (0x10b4, 0x34), // USB3_DP_QSERDES_COM_LOCK_CMP1_MODE1
-    (0x10b8, 0x82), // USB3_DP_QSERDES_COM_LOCK_CMP2_MODE1
-    (0x10bc, 0x82), // USB3_DP_QSERDES_COM_DEC_START_MODE0
-    (0x10c4, 0x82), // USB3_DP_QSERDES_COM_DEC_START_MODE1
-    (0x10cc, 0xab), // USB3_DP_QSERDES_COM_DIV_FRAC_START1_MODE0
-    (0x10d0, 0xea), // USB3_DP_QSERDES_COM_DIV_FRAC_START2_MODE0
-    (0x10d4, 0x02), // USB3_DP_QSERDES_COM_DIV_FRAC_START3_MODE0
-    (0x10d8, 0xab), // USB3_DP_QSERDES_COM_DIV_FRAC_START1_MODE1
-    (0x10dc, 0xea), // USB3_DP_QSERDES_COM_DIV_FRAC_START2_MODE1
-    (0x10e0, 0x02), // USB3_DP_QSERDES_COM_DIV_FRAC_START3_MODE1
-    (0x110c, 0x02), // USB3_DP_QSERDES_COM_VCO_TUNE_MAP
-    (0x1110, 0x24), // USB3_DP_QSERDES_COM_VCO_TUNE1_MODE0
-    (0x1118, 0x24), // USB3_DP_QSERDES_COM_VCO_TUNE1_MODE1
-    (0x111c, 0x02), // USB3_DP_QSERDES_COM_VCO_TUNE2_MODE1
-    (0x1158, 0x01), // USB3_DP_QSERDES_COM_HSCLK_SEL
-    (0x116c, 0x08), // USB3_DP_QSERDES_COM_CORECLK_DIV_MODE1
-    (0x11ac, 0xca), // USB3_DP_QSERDES_COM_BIN_VCOCAL_CMP_CODE1_MODE0
-    (0x11b0, 0x1e), // USB3_DP_QSERDES_COM_BIN_VCOCAL_CMP_CODE2_MODE0
-    (0x11b4, 0xca), // USB3_DP_QSERDES_COM_BIN_VCOCAL_CMP_CODE1_MODE1
-    (0x11b8, 0x1e), // USB3_DP_QSERDES_COM_BIN_VCOCAL_CMP_CODE2_MODE1
-    (0x11bc, 0x11), // USB3_DP_QSERDES_COM_BIN_VCOCAL_HSCLK_SEL
-    (0x1234, 0x00), // USB3_DP_QSERDES_TXA_RES_CODE_LANE_TX
-    (0x1238, 0x00), // USB3_DP_QSERDES_TXA_RES_CODE_LANE_RX
-    (0x123c, 0x16), // USB3_DP_QSERDES_TXA_RES_CODE_LANE_OFFSET_TX
-    (0x1240, 0x05), // USB3_DP_QSERDES_TXA_RES_CODE_LANE_OFFSET_RX
-    (0x1284, 0x55), // USB3_DP_QSERDES_TXA_LANE_MODE_1
-    (0x1288, 0x02), // USB3_DP_QSERDES_TXA_LANE_MODE_2
-    (0x1290, 0x2a), // USB3_DP_QSERDES_TXA_LANE_MODE_4
-    (0x1294, 0x3f), // USB3_DP_QSERDES_TXA_LANE_MODE_5
-    (0x12a4, 0x12), // USB3_DP_QSERDES_TXA_RCV_DETECT_LVL_2
-    (0x12e4, 0x20), // USB3_DP_QSERDES_TXA_PI_QEC_CTRL
-    (0x1414, 0x05), // USB3_DP_QSERDES_RXA_UCDR_SO_GAIN
-    (0x1430, 0x2f), // USB3_DP_QSERDES_RXA_UCDR_FASTLOCK_FO_GAIN
-    (0x1434, 0x7f), // USB3_DP_QSERDES_RXA_UCDR_SO_SATURATION_AND_ENABLE
-    (0x143c, 0xff), // USB3_DP_QSERDES_RXA_UCDR_FASTLOCK_COUNT_LOW
-    (0x1440, 0x0f), // USB3_DP_QSERDES_RXA_UCDR_FASTLOCK_COUNT_HIGH
-    (0x1444, 0x99), // USB3_DP_QSERDES_RXA_UCDR_PI_CONTROLS
-    (0x144c, 0x04), // USB3_DP_QSERDES_RXA_UCDR_SB2_THRESH1
-    (0x1450, 0x08), // USB3_DP_QSERDES_RXA_UCDR_SB2_THRESH2
-    (0x1454, 0x05), // USB3_DP_QSERDES_RXA_UCDR_SB2_GAIN1
-    (0x1458, 0x05), // USB3_DP_QSERDES_RXA_UCDR_SB2_GAIN2
-    (0x14d4, 0x54), // USB3_DP_QSERDES_RXA_VGA_CAL_CNTRL1
-    (0x14d8, 0x08), // USB3_DP_QSERDES_RXA_VGA_CAL_CNTRL2
-    (0x14ec, 0x0f), // USB3_DP_QSERDES_RXA_RX_EQU_ADAPTOR_CNTRL2
-    (0x14f0, 0x4a), // USB3_DP_QSERDES_RXA_RX_EQU_ADAPTOR_CNTRL3
-    (0x14f4, 0x0a), // USB3_DP_QSERDES_RXA_RX_EQU_ADAPTOR_CNTRL4
-    (0x14f8, 0xc0), // USB3_DP_QSERDES_RXA_RX_IDAC_TSETTLE_LOW
-    (0x14fc, 0x00), // USB3_DP_QSERDES_RXA_RX_IDAC_TSETTLE_HIGH
-    (0x1510, 0x77), // USB3_DP_QSERDES_RXA_RX_EQ_OFFSET_ADAPTOR_CNTRL1
-    (0x151c, 0x04), // USB3_DP_QSERDES_RXA_SIGDET_CNTRL
-    (0x1524, 0x0e), // USB3_DP_QSERDES_RXA_SIGDET_DEGLITCH_CNTRL
-    (0x155c, 0xbf), // USB3_DP_QSERDES_RXA_RX_MODE_00_LOW
-    (0x1560, 0xbf), // USB3_DP_QSERDES_RXA_RX_MODE_00_HIGH
-    (0x1564, 0x3f), // USB3_DP_QSERDES_RXA_RX_MODE_00_HIGH2
-    (0x1568, 0x7f), // USB3_DP_QSERDES_RXA_RX_MODE_00_HIGH3
-    (0x156c, 0x94), // USB3_DP_QSERDES_RXA_RX_MODE_00_HIGH4
-    (0x1570, 0x5b), // USB3_DP_QSERDES_RXA_RX_MODE_01_LOW
-    (0x1574, 0x1b), // USB3_DP_QSERDES_RXA_RX_MODE_01_HIGH
-    (0x1578, 0xd2), // USB3_DP_QSERDES_RXA_RX_MODE_01_HIGH2
-    (0x157c, 0x13), // USB3_DP_QSERDES_RXA_RX_MODE_01_HIGH3
-    (0x1580, 0xa9), // USB3_DP_QSERDES_RXA_RX_MODE_01_HIGH4
-    (0x15a0, 0x04), // USB3_DP_QSERDES_RXA_DFE_EN_TIMER
-    (0x15a4, 0x00), // USB3_DP_QSERDES_RXA_DFE_CTLE_POST_CAL_OFFSET
-    (0x1460, 0xa0), // USB3_DP_QSERDES_RXA_AUX_DATA_TCOARSE_TFINE
-    (0x15a8, 0x0c), // USB3_DP_QSERDES_RXA_DCC_CTRL1
-    (0x14dc, 0x00), // USB3_DP_QSERDES_RXA_GM_CAL
-    (0x15b0, 0x10), // USB3_DP_QSERDES_RXA_VTH_CODE
-    (0x1634, 0x00), // USB3_DP_QSERDES_TXB_RES_CODE_LANE_TX
-    (0x1638, 0x00), // USB3_DP_QSERDES_TXB_RES_CODE_LANE_RX
-    (0x163c, 0x16), // USB3_DP_QSERDES_TXB_RES_CODE_LANE_OFFSET_TX
-    (0x1640, 0x05), // USB3_DP_QSERDES_TXB_RES_CODE_LANE_OFFSET_RX
-    (0x1684, 0x55), // USB3_DP_QSERDES_TXB_LANE_MODE_1
-    (0x1688, 0x02), // USB3_DP_QSERDES_TXB_LANE_MODE_2
-    (0x1690, 0x2a), // USB3_DP_QSERDES_TXB_LANE_MODE_4
-    (0x1694, 0x3f), // USB3_DP_QSERDES_TXB_LANE_MODE_5
-    (0x16a4, 0x12), // USB3_DP_QSERDES_TXB_RCV_DETECT_LVL_2
-    (0x16e4, 0x02), // USB3_DP_QSERDES_TXB_PI_QEC_CTRL
-    (0x1814, 0x05), // USB3_DP_QSERDES_RXB_UCDR_SO_GAIN
-    (0x1830, 0x2f), // USB3_DP_QSERDES_RXB_UCDR_FASTLOCK_FO_GAIN
-    (0x1834, 0x7f), // USB3_DP_QSERDES_RXB_UCDR_SO_SATURATION_AND_ENABLE
-    (0x183c, 0xff), // USB3_DP_QSERDES_RXB_UCDR_FASTLOCK_COUNT_LOW
-    (0x1840, 0x0f), // USB3_DP_QSERDES_RXB_UCDR_FASTLOCK_COUNT_HIGH
-    (0x1844, 0x99), // USB3_DP_QSERDES_RXB_UCDR_PI_CONTROLS
-    (0x184c, 0x04), // USB3_DP_QSERDES_RXB_UCDR_SB2_THRESH1
-    (0x1850, 0x08), // USB3_DP_QSERDES_RXB_UCDR_SB2_THRESH2
-    (0x1854, 0x05), // USB3_DP_QSERDES_RXB_UCDR_SB2_GAIN1
-    (0x1858, 0x05), // USB3_DP_QSERDES_RXB_UCDR_SB2_GAIN2
-    (0x18d4, 0x54), // USB3_DP_QSERDES_RXB_VGA_CAL_CNTRL1
-    (0x18d8, 0x08), // USB3_DP_QSERDES_RXB_VGA_CAL_CNTRL2
-    (0x18ec, 0x0f), // USB3_DP_QSERDES_RXB_RX_EQU_ADAPTOR_CNTRL2
-    (0x18f0, 0x4a), // USB3_DP_QSERDES_RXB_RX_EQU_ADAPTOR_CNTRL3
-    (0x18f4, 0x0a), // USB3_DP_QSERDES_RXB_RX_EQU_ADAPTOR_CNTRL4
-    (0x18f8, 0xc0), // USB3_DP_QSERDES_RXB_RX_IDAC_TSETTLE_LOW
-    (0x18fc, 0x00), // USB3_DP_QSERDES_RXB_RX_IDAC_TSETTLE_HIGH
-    (0x1910, 0x77), // USB3_DP_QSERDES_RXB_RX_EQ_OFFSET_ADAPTOR_CNTRL1
-    (0x191c, 0x04), // USB3_DP_QSERDES_RXB_SIGDET_CNTRL
-    (0x1924, 0x0e), // USB3_DP_QSERDES_RXB_SIGDET_DEGLITCH_CNTRL
-    (0x195c, 0xbf), // USB3_DP_QSERDES_RXB_RX_MODE_00_LOW
-    (0x1960, 0xbf), // USB3_DP_QSERDES_RXB_RX_MODE_00_HIGH
-    (0x1964, 0x3f), // USB3_DP_QSERDES_RXB_RX_MODE_00_HIGH2
-    (0x1968, 0x7f), // USB3_DP_QSERDES_RXB_RX_MODE_00_HIGH3
-    (0x196c, 0x94), // USB3_DP_QSERDES_RXB_RX_MODE_00_HIGH4
-    (0x1970, 0x5b), // USB3_DP_QSERDES_RXB_RX_MODE_01_LOW
-    (0x1974, 0x1b), // USB3_DP_QSERDES_RXB_RX_MODE_01_HIGH
-    (0x1978, 0xd2), // USB3_DP_QSERDES_RXB_RX_MODE_01_HIGH2
-    (0x197c, 0x13), // USB3_DP_QSERDES_RXB_RX_MODE_01_HIGH3
-    (0x1980, 0xa9), // USB3_DP_QSERDES_RXB_RX_MODE_01_HIGH4
-    (0x19a0, 0x04), // USB3_DP_QSERDES_RXB_DFE_EN_TIMER
-    (0x19a4, 0x00), // USB3_DP_QSERDES_RXB_DFE_CTLE_POST_CAL_OFFSET
-    (0x1860, 0xa0), // USB3_DP_QSERDES_RXB_AUX_DATA_TCOARSE_TFINE
-    (0x19a8, 0x0c), // USB3_DP_QSERDES_RXB_DCC_CTRL1
-    (0x18dc, 0x00), // USB3_DP_QSERDES_RXB_GM_CAL
-    (0x19b0, 0x10), // USB3_DP_QSERDES_RXB_VTH_CODE
-    (0x1cc4, 0xd0), // USB3_DP_PCS_LOCK_DETECT_CONFIG1
-    (0x1cc8, 0x07), // USB3_DP_PCS_LOCK_DETECT_CONFIG2
-    (0x1ccc, 0x20), // USB3_DP_PCS_LOCK_DETECT_CONFIG3
-    (0x1cd8, 0x13), // USB3_DP_PCS_LOCK_DETECT_CONFIG6
-    (0x1cdc, 0x21), // USB3_DP_PCS_REFGEN_REQ_CONFIG1
-    (0x1d88, 0xaa), // USB3_DP_PCS_RX_SIGDET_LVL
-    (0x1db0, 0x0f), // USB3_DP_PCS_CDR_RESET_TIME
-    (0x1dc0, 0x88), // USB3_DP_PCS_ALIGN_DETECT_CONFIG1
-    (0x1dc4, 0x13), // USB3_DP_PCS_ALIGN_DETECT_CONFIG2
-    (0x1dd0, 0x0c), // USB3_DP_PCS_PCS_TX_RX_CONFIG
-    (0x1ddc, 0x4b), // USB3_DP_PCS_EQ_CONFIG1
-    (0x1dec, 0x10), // USB3_DP_PCS_EQ_CONFIG5
-    (0x1f18, 0xf8), // USB3_DP_PCS_USB3_LFPS_DET_HIGH_COUNT_VAL
-    (0x1f38, 0x07), // USB3_DP_PCS_USB3_RXEQTRAINING_DFE_TIME_S2
-];
-
-/// Active PHY tables. The compiled values are the Bramble fallback, while
-/// the DT path may replace them after validating the complete vendor
-/// property. Keeping the delay array separate preserves the compact static
-/// table and still executes the DT's third cell rather than silently dropping
-/// it.
-static mut ACTIVE_QMP_INIT: [(usize, u32); 146] = QMP_INIT;
-static mut ACTIVE_QMP_INIT_DELAY_US: [u32; 146] = [0; 146];
-static mut ACTIVE_HSPHY_PARAM_OVERRIDE: [(usize, u32); 3] =
-    [(0x6c, 0x63), (0x70, 0x85), (0x74, 0x17)];
-
-/// Install the complete PHY programming properties from the bootloader DTB.
-/// A partial or malformed property is rejected as a unit, leaving the known
-/// Bramble fallback in place. The QMP binding terminates its 146 triples with
-/// `<0xffffffff 0xffffffff 0>`, which is a sentinel and is not written.
-pub fn install_dt_phy_sequences(hs_raw: [Option<u32>; 6], qmp_raw: [Option<u32>; 441]) -> bool {
-    let mut installed = false;
-
-    if hs_raw.iter().all(Option::is_some) {
-        let mut entries = [(0usize, 0u32); 3];
-        let mut valid = true;
-        for index in 0..3 {
-            let value = hs_raw[index * 2].unwrap();
-            let offset = hs_raw[index * 2 + 1].unwrap();
-            valid &= value <= 0xff
-                && matches!(offset, 0x6c | 0x70 | 0x74)
-                && entries[..index]
-                    .iter()
-                    .all(|entry| entry.0 != offset as usize);
-            entries[index] = (offset as usize, value);
-        }
-        if valid {
-            unsafe { ACTIVE_HSPHY_PARAM_OVERRIDE = entries };
-            installed = true;
-        }
-    }
-
-    if qmp_raw.iter().all(Option::is_some)
-        && qmp_raw[438] == Some(u32::MAX)
-        && qmp_raw[439] == Some(u32::MAX)
-        && qmp_raw[440] == Some(0)
-    {
-        let mut entries = [(0usize, 0u32); 146];
-        let mut delays = [0u32; 146];
-        let mut valid = true;
-        for index in 0..146 {
-            let raw = index * 3;
-            let offset = qmp_raw[raw].unwrap();
-            let value = qmp_raw[raw + 1].unwrap();
-            let delay_us = qmp_raw[raw + 2].unwrap();
-            valid &= offset <= 0x2fff && value <= 0xff && delay_us <= 1_000_000;
-            entries[index] = (offset as usize, value);
-            delays[index] = delay_us;
-        }
-        if valid {
-            unsafe {
-                ACTIVE_QMP_INIT = entries;
-                ACTIVE_QMP_INIT_DELAY_US = delays;
-            }
-            installed = true;
-        }
-    }
-
-    installed
 }
 
 const EVENT_BUFFER_SIZE: usize = 4096;
@@ -1063,339 +229,11 @@ unsafe fn dma_iova_for(cpu: usize) -> u64 {
     }
 }
 
-/// Read-only Apps-SMMU stage-1 walk. Returns the physical output address for
-/// `iova` when the stream's context bank translation maps it with a valid
-/// 4 KiB page (or 2 MiB block whose base contains the whole 4 KiB window).
-unsafe fn smmu_walk_iova(ttbr0: u64, three_level: bool, iova: u64) -> Option<u64> {
-    const DESC_VALID: u64 = 1;
-    const DESC_TABLE: u64 = 2;
-    unsafe {
-        let mut table = (ttbr0 & !0xfff) as usize;
-        // 39-bit IOVA (T0SZ=25, 4 KiB granule): L1 root. 32-bit IOVA
-        // (T0SZ=32): the walk starts at level 2 with 2 MiB blocks.
-        let (l1_index, l2_index, l3_index) = if three_level {
-            (
-                ((iova >> 30) & 0x1ff) as usize,
-                ((iova >> 21) & 0x1ff) as usize,
-                ((iova >> 12) & 0x1ff) as usize,
-            )
-        } else {
-            (
-                0,
-                ((iova >> 21) & 0x1ff) as usize,
-                ((iova >> 12) & 0x1ff) as usize,
-            )
-        };
-        if three_level {
-            let descriptor = read_volatile((table + l1_index * 8) as *const u64);
-            if descriptor & DESC_VALID == 0 || descriptor & DESC_TABLE == 0 {
-                return None;
-            }
-            table = (descriptor & !0xfff) as usize;
-        }
-        let l2_descriptor = read_volatile((table + l2_index * 8) as *const u64);
-        if l2_descriptor & DESC_VALID == 0 {
-            return None;
-        }
-        if l2_descriptor & DESC_TABLE == 0 {
-            // 2 MiB block: the whole 4 KiB window must fit inside it. The
-            // block base is 2 MiB aligned, so the low attribute bits do not
-            // overlap the output address.
-            let block = (l2_descriptor & 0x000f_ffff_ffff_f000) as usize;
-            let offset = (iova as usize) & 0x1f_ffff;
-            if offset + 0x1000 > 0x20_0000 {
-                return None;
-            }
-            return Some((block | offset) as u64);
-        }
-        let l3_table = (l2_descriptor & !0xfff) as usize;
-        let l3_descriptor = read_volatile((l3_table + l3_index * 8) as *const u64);
-        if l3_descriptor & DESC_VALID == 0 || l3_descriptor & DESC_TABLE != 0 {
-            return None;
-        }
-        Some(l3_descriptor & 0x000f_ffff_ffff_f000)
-    }
-}
-
-/// Read-only Apps-SMMU stream/context discovery. Returns the context-bank
-/// page index when the DWC3 stream matches a valid SMR whose S2CR selects an
-/// active TRANSLATE context. Never writes the SMMU.
-unsafe fn smmu_find_translate_context() -> Option<(usize, usize, usize)> {
-    unsafe {
-        let id0 = read_volatile(smmu_reg(SMMU_ID0));
-        let id1 = read_volatile(smmu_reg(SMMU_ID1));
-        if id0 == 0 || id0 == u32::MAX || id1 == 0 || id1 == u32::MAX {
-            return None;
-        }
-        let num_smrs = ((id0 & SMMU_ID0_NUMSMRG_MASK) as usize).min(128);
-        let num_pages =
-            1usize << (((id1 >> SMMU_ID1_NUMPAGENDXB_SHIFT) & SMMU_ID1_NUMPAGENDXB_MASK) + 1);
-        let num_context_banks = (id1 & SMMU_ID1_NUMCB_MASK) as usize;
-        if num_pages == 0 || num_context_banks == 0 {
-            return None;
-        }
-        let stream_id = super::platform::bramble::usb_resources().dma_pool.stream_id;
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID == 0 {
-                continue;
-            }
-            let id = smr & 0xffff;
-            let mask = (smr >> SMMU_SMR_MASK_SHIFT) & 0x7fff;
-            if (stream_id ^ id) & !mask != 0 {
-                continue;
-            }
-            let s2cr = read_volatile(smmu_reg(SMMU_S2CR_BASE + index * 4));
-            if s2cr & SMMU_S2CR_TYPE_MASK != SMMU_S2CR_TYPE_TRANS {
-                // BYPASS or FAULT: the CPU address is already the DMA
-                // address and the normal linker section works.
-                return None;
-            }
-            let cbndx = (s2cr & SMMU_S2CR_CBNDX_MASK) as usize;
-            if cbndx >= num_context_banks {
-                return None;
-            }
-            return Some((cbndx, num_pages, num_context_banks));
-        }
-        None
-    }
-}
-
-/// Find ANY valid 4 KiB page mapping inside the live context's stage-1 tables
-/// and return (iova, physical). The bootloader context is no longer active
-/// once `fastboot boot` jumped away, so any page it mapped is a fair DMA
-/// window for Fullerene's EP0 objects.
-unsafe fn smmu_find_any_mapping(ttbr0: u64, three_level: bool) -> Option<(u64, u64)> {
-    const DESC_VALID: u64 = 1;
-    const DESC_TABLE: u64 = 2;
-    unsafe {
-        let l1_table = (ttbr0 & !0xfff) as usize;
-        let l1_range = if three_level { 512usize } else { 1 };
-        for l1_index in 0..l1_range {
-            let mut l2_table = l1_table;
-            if three_level {
-                let l1_descriptor = read_volatile((l1_table + l1_index * 8) as *const u64);
-                if l1_descriptor & DESC_VALID == 0 || l1_descriptor & DESC_TABLE == 0 {
-                    continue;
-                }
-                l2_table = (l1_descriptor & !0xfff) as usize;
-            }
-            for l2_index in 0..512usize {
-                let l2_descriptor = read_volatile((l2_table + l2_index * 8) as *const u64);
-                if l2_descriptor & DESC_VALID == 0 {
-                    continue;
-                }
-                let iova_base = ((l1_index as u64) << 30) | ((l2_index as u64) << 21);
-                if l2_descriptor & DESC_TABLE == 0 {
-                    // A 2 MiB block: use its first 4 KiB window.
-                    let block = l2_descriptor & 0x000f_ffff_ffff_f000;
-                    return Some((iova_base, block));
-                }
-                let l3_table = (l2_descriptor & !0xfff) as usize;
-                for l3_index in 0..512usize {
-                    let l3_descriptor = read_volatile((l3_table + l3_index * 8) as *const u64);
-                    if l3_descriptor & DESC_VALID == 0 || l3_descriptor & DESC_TABLE != 0 {
-                        continue;
-                    }
-                    let physical = l3_descriptor & 0x000f_ffff_ffff_f000;
-                    if physical == 0 {
-                        continue;
-                    }
-                    return Some((iova_base | ((l3_index as u64) << 12), physical));
-                }
-            }
-        }
-        None
-    }
-}
-
 /// Newest STARTTRANSFER outcome harvested from the retained trace of the
 /// previous attempts (0xFFFF_FFFF = none; bit 16 set = the command timed out;
 /// otherwise the raw DEPCMD register: status in bits 15:12).
 pub fn harvest_last_str_code() -> u32 {
     unsafe { TRACE_HARVEST_LAST }
-}
-
-/// Read-only Apps-SMMU stream-state ladder. Returns the deepest condition
-/// that provably holds, so a host-visible attach gate can name the state one
-/// run at a time:
-///   0..=3 = an SMR matched the stream and its S2CR type is that value
-///   251   = SMRs are implemented but none is valid
-///   252   = at least one valid SMR exists but none matches the stream
-///   253   = no SMRs are implemented (ID0.NUMSMRG == 0)
-///   254   = the SMMU identification registers are unreadable (RAZ/all-ones)
-unsafe fn smmu_stream_s2cr_type() -> u32 {
-    unsafe {
-        let id0 = read_volatile(smmu_reg(SMMU_ID0));
-        let id1 = read_volatile(smmu_reg(SMMU_ID1));
-        if id0 == 0 || id0 == u32::MAX || id1 == 0 || id1 == u32::MAX {
-            return 254;
-        }
-        let num_smrs = ((id0 & SMMU_ID0_NUMSMRG_MASK) as usize).min(128);
-        if num_smrs == 0 {
-            return 253;
-        }
-        let stream_id = super::platform::bramble::usb_resources().dma_pool.stream_id;
-        let mut any_valid = false;
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID == 0 {
-                continue;
-            }
-            any_valid = true;
-            let id = smr & 0xffff;
-            let mask = (smr >> SMMU_SMR_MASK_SHIFT) & 0x7fff;
-            if (stream_id ^ id) & !mask != 0 {
-                continue;
-            }
-            let s2cr = read_volatile(smmu_reg(SMMU_S2CR_BASE + index * 4));
-            return (s2cr & SMMU_S2CR_TYPE_MASK) >> 16;
-        }
-        if any_valid {
-            // Stream unmatched. Distinguish an active SMMU (252) from a
-            // globally bypassed one (250): with SMMUEN=0 or CLIENTPD=1 the
-            // unmatched SMR is irrelevant and transactions already pass
-            // untranslated.
-            let scr0 = read_volatile(smmu_reg(SMMU_GR0_SCR0));
-            let active = scr0 != u32::MAX && (scr0 & 1) != 0 && (scr0 & 2) == 0;
-            return if active { 252 } else { 250 };
-        }
-        251
-    }
-}
-
-/// Claim a free Apps-SMMU SMR slot for the DWC3 stream and point its S2CR at
-/// BYPASS so the stream's transactions pass untranslated (CPU address ==
-/// DMA address). Only slots whose VALID bit is clear are claimed, both
-/// writes are verified by readback, and a rejected (secure-owned) write is
-/// reported as failure instead of being assumed. The stream is known to be
-/// unmatched (ladder 252) when this runs, so no live mapping is displaced.
-unsafe fn smmu_install_stream_bypass() -> bool {
-    unsafe {
-        let id0 = read_volatile(smmu_reg(SMMU_ID0));
-        let id1 = read_volatile(smmu_reg(SMMU_ID1));
-        if id0 == 0 || id0 == u32::MAX || id1 == 0 || id1 == u32::MAX {
-            return false;
-        }
-        let num_smrs = ((id0 & SMMU_ID0_NUMSMRG_MASK) as usize).min(128);
-        if num_smrs == 0 {
-            return false;
-        }
-        let stream_id = super::platform::bramble::usb_resources().dma_pool.stream_id & 0xffff;
-        // An existing match means the stream already has an owner: leave the
-        // configuration alone and report success (nothing to install).
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID == 0 {
-                continue;
-            }
-            let id = smr & 0xffff;
-            let mask = (smr >> SMMU_SMR_MASK_SHIFT) & 0x7fff;
-            if (stream_id ^ id) & !mask == 0 {
-                trace_event(TRACE_SMMU_HANDOFF, 0x494E, index as u32, 0, 1, 0);
-                return true;
-            }
-        }
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID != 0 {
-                continue;
-            }
-            // Claim this slot. Write the inert S2CR first (it only applies
-            // once SMR.VALID is set), then publish the SMR, then verify both
-            // by readback before declaring the stream owned. The catch-all
-            // mode matches every stream ID so a misreported DWC3 stream ID
-            // cannot keep the stream faulting.
-            let catch_all = option_env!("FULLERENE_USB_SMMU_INSTALL_ALL") == Some("1");
-            let smr_value = if catch_all {
-                SMMU_SMR_VALID | (0x7fff << SMMU_SMR_MASK_SHIFT)
-            } else {
-                SMMU_SMR_VALID | stream_id
-            };
-            let s2cr_address = SMMU_S2CR_BASE + index * 4;
-            let old_s2cr = read_volatile(smmu_reg(s2cr_address));
-            let new_s2cr = (old_s2cr & !SMMU_S2CR_TYPE_MASK) | SMMU_S2CR_TYPE_BYPASS;
-            write_volatile(smmu_reg(s2cr_address), new_s2cr);
-            core::arch::asm!("dsb sy", options(nostack));
-            write_volatile(smmu_reg(SMMU_SMR_BASE + index * 4), smr_value);
-            core::arch::asm!("dsb sy", options(nostack));
-            let smr_readback = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            let s2cr_readback = read_volatile(smmu_reg(s2cr_address));
-            let smr_ok = smr_readback & SMMU_SMR_VALID != 0 && smr_readback == smr_value;
-            let s2cr_ok = s2cr_readback & SMMU_S2CR_TYPE_MASK == SMMU_S2CR_TYPE_BYPASS;
-            trace_event(
-                TRACE_SMMU_HANDOFF,
-                0x534D_5200 | index as u32,
-                smr_readback,
-                s2cr_readback,
-                smr_ok as u32,
-                s2cr_ok as u32,
-            );
-            if smr_ok && s2cr_ok {
-                smmu_tlb_sync();
-                log_hex("usb: installed SMMU bypass SMR index=", index as u64);
-                return true;
-            }
-            // The secure side rejected at least one write. Restore the slot
-            // to its inert state as far as non-secure writes reach and
-            // report the failure.
-            write_volatile(smmu_reg(SMMU_SMR_BASE + index * 4), smr);
-            write_volatile(smmu_reg(s2cr_address), old_s2cr);
-            core::arch::asm!("dsb sy", options(nostack));
-            return false;
-        }
-        false
-    }
-}
-/// Relocate the EP0 DMA objects into a page that the live Apps-SMMU context
-/// already maps. The stream's S2CR is TRANSLATE and software cannot rewrite
-/// it from non-secure state, so the only working DMA window is one the
-/// bootloader context maps: capture the bootloader's event-ring IOVA, walk
-/// its page tables read-only, and adopt that page (CPU side = physical,
-/// DWC3 side = IOVA). Returns the IOVA of the adopted page.
-unsafe fn adopt_smmu_dma_mapping() -> Option<u64> {
-    unsafe {
-        let (cbndx, num_pages, _) = smmu_find_translate_context()?;
-        let page_size = 0x1000usize;
-        let cb_page = num_pages + cbndx;
-        let sctlr = smmu_page_read(page_size, cb_page, SMMU_CB_SCTLR);
-        if sctlr & SMMU_SCTLR_M == 0 {
-            // Translation disabled: the CPU address is the DMA address.
-            return None;
-        }
-        let ttbr0 = smmu_page_read64(page_size, cb_page, SMMU_CB_TTBR0);
-        if ttbr0 == 0 || ttbr0 == u64::MAX {
-            return None;
-        }
-        let three_level = super::platform::bramble::usb_resources().smmu_use_3_level_tables;
-        let iova =
-            (read_volatile(reg(GEVNTADRHI0)) as u64) << 32 | read_volatile(reg(GEVNTADRLO0)) as u64;
-        if iova == 0
-            || iova == u64::MAX
-            || iova & 0xfff != 0
-            || iova > usize::MAX as u64
-            || iova >= 1 << (if three_level { 39 } else { 32 })
-        {
-            return None;
-        }
-        let (iova, physical) = match smmu_walk_iova(ttbr0, three_level, iova) {
-            Some(physical) if physical != 0 => (iova, physical as u64),
-            // The bootloader's own event-ring IOVA is no longer resolvable
-            // (its teardown may have cleared GEVNTADR or the page). Adopt
-            // ANY page the live context still maps instead: the bootloader
-            // is gone and no other master owns the stream.
-            _ => smmu_find_any_mapping(ttbr0, three_level)?,
-        };
-        if physical == 0 || iova & 0xfff != 0 {
-            return None;
-        }
-        DMA_ADOPTED_CPU = physical as usize;
-        DMA_ADOPTED_IOVA = iova;
-        DMA_ADOPTED = true;
-        log_hex("usb: adopted SMMU page physical=", physical);
-        log_hex("usb: adopted SMMU page iova=", iova);
-        Some(iova)
-    }
 }
 
 #[inline]
@@ -1627,80 +465,6 @@ unsafe fn unbind_function() {
     }
 }
 
-const USB_TRACE_CAPACITY: usize = 256;
-
-// Numeric events keep the early USB path independent of UART, locks, and
-// formatting. The buffer is CPU-owned; it is placed beside the DMA objects so
-// a probe can preserve the same identity-mapped address discipline.
-const TRACE_INIT: u32 = 1;
-const TRACE_DEVICE_RESET: u32 = 2;
-const TRACE_DEVICE_CONNECT: u32 = 3;
-const TRACE_EP_COMMAND_ISSUE: u32 = 4;
-const TRACE_EP_COMMAND_DONE: u32 = 5;
-const TRACE_EP_COMMAND_TIMEOUT: u32 = 6;
-const TRACE_SETUP_QUEUED: u32 = 7;
-const TRACE_SETUP_RECEIVED: u32 = 8;
-const TRACE_DESCRIPTOR_QUEUED: u32 = 9;
-const TRACE_STATUS_QUEUED: u32 = 10;
-const TRACE_TRANSFER_COMPLETE: u32 = 11;
-const TRACE_USB_RESET: u32 = 12;
-pub const TRACE_BOOT_USB_ENTRY: u32 = 13;
-pub const TRACE_TYPEC_BEGIN: u32 = 14;
-pub const TRACE_TYPEC_DONE: u32 = 15;
-pub const TRACE_USB_HANDOFF_BEGIN: u32 = 16;
-const TRACE_DWC3_RESET_BEGIN: u32 = 17;
-const TRACE_QSCRATCH_BEGIN: u32 = 18;
-pub const TRACE_EXCEPTION_SYNC: u32 = 19;
-pub const TRACE_PROBE_WATCHDOG: u32 = 33;
-const TRACE_LINK_STATUS: u32 = 20;
-const TRACE_USB_WAKEUP: u32 = 21;
-const TRACE_USB_SUSPEND: u32 = 22;
-const TRACE_USB_DEVICE_ERROR: u32 = 23;
-pub const TRACE_TYPEC_EVENT: u32 = 24;
-pub const TRACE_PLATFORM_IRQ: u32 = 25;
-pub const TRACE_UDC_REARM: u32 = 26;
-const TRACE_SMMU_BEGIN: u32 = 27;
-const TRACE_SMMU_READY: u32 = 28;
-const TRACE_SMMU_HANDOFF: u32 = 34;
-const TRACE_SMMU_PRESERVED: u32 = 35;
-const TRACE_SMMU_FAULT: u32 = 36;
-const TRACE_SMMU_GLOBAL_FAULT: u32 = 37;
-const TRACE_UTMI_CLOCK: u32 = 29;
-const TRACE_EVENT_RING_READY: u32 = 30;
-const TRACE_DWC3_HALTED: u32 = 31;
-const TRACE_DWC3_HALT_TIMEOUT: u32 = 32;
-const TRACE_DWC3_REVISION_QUIRK: u32 = 38;
-const TRACE_XFER_NOT_READY: u32 = 40;
-const TRACE_GCC_UTMI_CLOCK: u32 = 41;
-const TRACE_USB2_PHY_RESET: u32 = 42;
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-struct UsbTraceEntry {
-    sequence: u32,
-    event: u32,
-    request: u32,
-    value: u32,
-    index: u32,
-    length: u32,
-    ep0_state: u32,
-    status: u32,
-}
-
-const EMPTY_USB_TRACE: UsbTraceEntry = UsbTraceEntry {
-    sequence: 0,
-    event: 0,
-    request: 0,
-    value: 0,
-    index: 0,
-    length: 0,
-    ep0_state: 0,
-    status: 0,
-};
-
-const USB_TRACE_MAGIC: u32 = 0x4655_5452; // "FUTR"
-const USB_TRACE_VERSION: u32 = 1;
-
 /// Outcome of the previous attempt's last STARTTRANSFER command, harvested
 /// from the retained trace at the start of the next handoff attempt (see
 /// `harvest_trace_outcome()`). Encoding: 0xFFFF = no record found,
@@ -1852,11 +616,11 @@ unsafe fn harvest_trace_outcome() {
         TRACE_HARVEST_ARMED = 0;
         TRACE_HARVEST_ARM_SEQ = 0xFFFF_FFFF;
         TRACE_HARVEST_SETUP_SEQ = 0xFFFF_FFFF;
-            TRACE_HARVEST_CONNECT = 0;
-            TRACE_HARVEST_ADDR = 0;
-            TRACE_HARVEST_ADDR2 = 0;
-            TRACE_HARVEST_DARM = 0xFFFF_FFFF;
-            TRACE_HARVEST_LAST_SETUP = 0xFFFF_FFFF;
+        TRACE_HARVEST_CONNECT = 0;
+        TRACE_HARVEST_ADDR = 0;
+        TRACE_HARVEST_ADDR2 = 0;
+        TRACE_HARVEST_DARM = 0xFFFF_FFFF;
+        TRACE_HARVEST_LAST_SETUP = 0xFFFF_FFFF;
         TRACE_HARVEST_EP1_XFER = 0xFFFF_FFFF;
         TRACE_HARVEST_EP1_NRDY = 0;
         for offset in 0..count {
@@ -2104,24 +868,6 @@ pub fn rescue_read64() -> u32 {
     }
 }
 
-#[repr(C, align(4096))]
-struct UsbTraceBuffer {
-    magic: u32,
-    version: u32,
-    head: u32,
-    reserved: u32,
-    entries: [UsbTraceEntry; USB_TRACE_CAPACITY],
-}
-
-#[unsafe(link_section = ".usb_trace")]
-static mut USB_TRACE: UsbTraceBuffer = UsbTraceBuffer {
-    magic: 0,
-    version: 0,
-    head: 0,
-    reserved: 0,
-    entries: [EMPTY_USB_TRACE; USB_TRACE_CAPACITY],
-};
-
 #[derive(Clone, Copy, PartialEq, Eq)]
 enum Ep0State {
     Setup,
@@ -2165,188 +911,6 @@ pub unsafe fn allocate_usb_dma(size: usize, alignment: usize) -> Option<*mut u8>
         allocator
             .allocate(size as u64, alignment as u64)
             .map(|address| address as usize as *mut u8)
-    }
-}
-
-/// Initialize the retained trace header and append a boot boundary marker.
-/// The entry array is intentionally not cleared, so a subsequent boot can
-/// inspect the last attempt after a warm reset.
-fn trace_begin() {
-    unsafe {
-        let magic = read_volatile(addr_of!(USB_TRACE).cast::<u32>());
-        let version = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(1));
-        if magic != USB_TRACE_MAGIC || version != USB_TRACE_VERSION {
-            write_volatile(addr_of_mut!(USB_TRACE).cast::<u32>(), USB_TRACE_MAGIC);
-            write_volatile(
-                addr_of_mut!(USB_TRACE).cast::<u32>().add(1),
-                USB_TRACE_VERSION,
-            );
-            write_volatile(addr_of_mut!(USB_TRACE).cast::<u32>().add(2), 0);
-            write_volatile(addr_of_mut!(USB_TRACE).cast::<u32>().add(3), 0);
-        }
-    }
-    trace_event(TRACE_BOOT_USB_ENTRY, 0, 0, 0, 0, 0);
-}
-
-#[inline(always)]
-fn ep0_state_code(state: Ep0State) -> u32 {
-    match state {
-        Ep0State::Setup => 1,
-        Ep0State::Data => 2,
-        Ep0State::Status => 3,
-    }
-}
-
-#[inline(always)]
-fn trace_event(event: u32, request: u32, value: u32, index: u32, length: u32, status: u32) {
-    unsafe {
-        let head_ptr = addr_of_mut!(USB_TRACE).cast::<u32>().add(2);
-        let head = read_volatile(head_ptr);
-        let slot = (head as usize) % USB_TRACE_CAPACITY;
-        let entry = UsbTraceEntry {
-            sequence: head.wrapping_add(1),
-            event,
-            request,
-            value,
-            index,
-            length,
-            ep0_state: ep0_state_code(EP0_STATE),
-            status,
-        };
-        write_volatile(
-            addr_of_mut!(USB_TRACE.entries)
-                .cast::<UsbTraceEntry>()
-                .add(slot),
-            entry,
-        );
-        write_volatile(head_ptr, head.wrapping_add(1));
-    }
-}
-
-/// Start a retained trace for the standalone handoff probe without clearing
-/// the previous attempt. A subsequent normal Fullerene boot can dump it over
-/// UART before starting a new trace.
-pub fn trace_probe_begin() {
-    trace_begin();
-}
-
-/// Reset the retained trace cursor for a fresh boot. The region survives
-/// warm resets by design, but between two `fastboot boot` runs Android
-/// scribbles DRAM unpredictably: a surviving header would make the in-boot
-/// harvest gates count the PREVIOUS run's records. The probe calls this once
-/// per boot, before the first handoff attempt, so attempts 2/3 still see
-/// attempt 1/2's records while cross-boot contamination is impossible.
-pub fn trace_reset_head_for_boot() {
-    unsafe {
-        write_volatile(addr_of_mut!(USB_TRACE).cast::<u32>().add(2), 0);
-        core::arch::asm!("dsb sy", options(nostack));
-    }
-}
-
-/// Add a marker without touching the controller. This is used around PMIC
-/// and platform transitions where the next MMIO access itself may abort.
-pub fn trace_marker(event: u32, status: u32) {
-    trace_event(event, 0, 0, 0, 0, status);
-}
-
-/// Read the retained trace cursor without changing the controller state.
-/// Standalone probes use this as a watchdog activity signal: an EP0/device
-/// event advances the cursor, while a completely absent USB session does not.
-pub fn trace_head() -> u32 {
-    unsafe { read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(2)) }
-}
-
-/// Read the last committed event from the retained trace without advancing
-/// it. The serial-string transport uses this pair as a compact host-visible
-/// snapshot when the gadget has enumerated but UART is unavailable.
-pub fn trace_last_event() -> u32 {
-    unsafe {
-        let magic = read_volatile(addr_of!(USB_TRACE).cast::<u32>());
-        let version = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(1));
-        let head = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(2));
-        if magic != USB_TRACE_MAGIC || version != USB_TRACE_VERSION || head == 0 {
-            return 0;
-        }
-        let slot = (head.wrapping_sub(1) as usize) % USB_TRACE_CAPACITY;
-        read_volatile(
-            addr_of!(USB_TRACE.entries)
-                .cast::<UsbTraceEntry>()
-                .add(slot),
-        )
-        .event
-    }
-}
-
-/// Fill one page of the retained trace for the vendor control request. The
-/// page order is oldest-to-newest within the valid window, so a host can read
-/// a consistent bounded snapshot without knowing the physical RAM address.
-/// A request for page zero returns the header even when the trace is empty;
-/// malformed or out-of-range pages are rejected by returning `None`.
-unsafe fn fill_trace_control_response(
-    response: &mut [u8],
-    requested_length: usize,
-    page: u16,
-) -> Option<usize> {
-    let requested_length = requested_length.min(response.len());
-    if requested_length == 0 {
-        return None;
-    }
-    unsafe {
-        let magic = read_volatile(addr_of!(USB_TRACE).cast::<u32>());
-        let version = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(1));
-        if magic != USB_TRACE_MAGIC || version != USB_TRACE_VERSION {
-            return None;
-        }
-        let head = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(2));
-        let valid = (head as usize).min(USB_TRACE_CAPACITY);
-        let page_start = (page as usize).checked_mul(TRACE_CONTROL_PAGE_ENTRIES)?;
-        if page_start > valid {
-            return None;
-        }
-
-        response[..requested_length].fill(0);
-        let mut write_u32 = |offset: usize, value: u32| {
-            if offset + 4 <= requested_length {
-                response[offset..offset + 4].copy_from_slice(&value.to_le_bytes());
-            }
-        };
-        write_u32(0, magic);
-        write_u32(4, version);
-        write_u32(8, head);
-        write_u32(12, valid as u32);
-        if requested_length <= TRACE_CONTROL_HEADER_BYTES {
-            return Some(requested_length);
-        }
-
-        let available = valid.saturating_sub(page_start);
-        let records = available
-            .min(TRACE_CONTROL_PAGE_ENTRIES)
-            .min((requested_length - TRACE_CONTROL_HEADER_BYTES) / TRACE_CONTROL_ENTRY_BYTES);
-        let oldest = (head as usize).saturating_sub(valid);
-        for index in 0..records {
-            let slot = (oldest + page_start + index) % USB_TRACE_CAPACITY;
-            let entry = read_volatile(
-                addr_of!(USB_TRACE.entries)
-                    .cast::<UsbTraceEntry>()
-                    .add(slot),
-            );
-            let values = [
-                entry.sequence,
-                entry.event,
-                entry.request,
-                entry.value,
-                entry.index,
-                entry.length,
-                entry.ep0_state,
-                entry.status,
-            ];
-            let base = TRACE_CONTROL_HEADER_BYTES + index * TRACE_CONTROL_ENTRY_BYTES;
-            for (word, value) in values.into_iter().enumerate() {
-                response[base + word * 4..base + word * 4 + 4]
-                    .copy_from_slice(&value.to_le_bytes());
-            }
-        }
-        Some(TRACE_CONTROL_HEADER_BYTES + records * TRACE_CONTROL_ENTRY_BYTES)
     }
 }
 
@@ -2487,926 +1051,6 @@ unsafe fn stop_after_gadget_handoff_stage(stage: u32) -> bool {
     true
 }
 
-/// Dump the post-mortem USB trace after the controller has reached a safe
-/// UART-visible stage. The hot path above never calls this or formats text.
-pub fn dump_trace() {
-    unsafe {
-        let magic = read_volatile(addr_of!(USB_TRACE).cast::<u32>());
-        let version = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(1));
-        if magic != USB_TRACE_MAGIC || version != USB_TRACE_VERSION {
-            uart::puts("usb trace: no retained record\n");
-            return;
-        }
-        let head = read_volatile(addr_of!(USB_TRACE).cast::<u32>().add(2));
-        let count = (head as usize).min(USB_TRACE_CAPACITY);
-        let start = (head as usize).saturating_sub(count);
-        uart::puts("usb trace begin\n");
-        for offset in 0..count {
-            let slot = (start + offset) % USB_TRACE_CAPACITY;
-            let entry = read_volatile(
-                addr_of!(USB_TRACE.entries)
-                    .cast::<UsbTraceEntry>()
-                    .add(slot),
-            );
-            uart::put_hex("usb trace event=", entry.event as u64);
-            uart::put_hex(" request=", entry.request as u64);
-            uart::put_hex(" value=", entry.value as u64);
-            uart::put_hex(" index=", entry.index as u64);
-            uart::put_hex(" length=", entry.length as u64);
-            uart::put_hex(" state=", entry.ep0_state as u64);
-            uart::put_hex(" status=", entry.status as u64);
-        }
-        uart::puts("usb trace end\n");
-    }
-}
-
-#[inline]
-fn reg(offset: usize) -> *mut u32 {
-    (dwc3_base() + offset) as *mut u32
-}
-
-#[inline]
-fn qscratch_reg(offset: usize) -> *mut u32 {
-    (qscratch_base() + offset) as *mut u32
-}
-
-#[inline]
-fn hsphy_reg(offset: usize) -> *mut u32 {
-    (hsphy_base() + offset) as *mut u32
-}
-
-#[inline]
-fn qmp_reg(offset: usize) -> *mut u32 {
-    (qmp_base() + offset) as *mut u32
-}
-
-#[inline]
-fn qmp_contract_offset(slot: usize, fallback: usize) -> usize {
-    let offset = super::platform::bramble::usb_resources().qmp_reg_offsets[slot];
-    if offset == 0xffff { fallback } else { offset }
-}
-
-#[inline]
-unsafe fn smmu_reg(offset: usize) -> *mut u32 {
-    (apps_smmu_base() + offset) as *mut u32
-}
-
-#[inline]
-unsafe fn smmu_page_reg(page_size: usize, page: usize, offset: usize) -> *mut u32 {
-    (apps_smmu_base() + page * page_size + offset) as *mut u32
-}
-
-#[inline]
-unsafe fn smmu_page_write(page_size: usize, page: usize, offset: usize, value: u32) {
-    unsafe { write_volatile(smmu_page_reg(page_size, page, offset), value) };
-}
-
-#[inline]
-unsafe fn smmu_page_read(page_size: usize, page: usize, offset: usize) -> u32 {
-    unsafe { read_volatile(smmu_page_reg(page_size, page, offset)) }
-}
-
-#[inline]
-unsafe fn smmu_page_write64(page_size: usize, page: usize, offset: usize, value: u64) {
-    unsafe { write_volatile(smmu_page_reg(page_size, page, offset).cast::<u64>(), value) };
-}
-
-#[inline]
-unsafe fn smmu_page_read64(page_size: usize, page: usize, offset: usize) -> u64 {
-    unsafe { read_volatile(smmu_page_reg(page_size, page, offset).cast::<u64>()) }
-}
-
-/// Verify that an already-live S1 context translates the complete linker DMA
-/// section as an identity map. A Fastboot handoff may safely preserve Linux's
-/// context only when its page tables cover every Fullerene TRB/event object;
-/// preserving an unrelated bootloader map would make the first DMA fault look
-/// like an EP0 protocol failure.
-unsafe fn smmu_context_maps_identity(ttbr0: u64, tcr: u32, start: usize, end: usize) -> bool {
-    let iova_start = start as u64;
-    let iova_end = end as u64;
-    if ttbr0 == 0
-        || ttbr0 == u64::MAX
-        || start >= end
-        || iova_end > (1u64 << 39)
-        || (tcr >> 14) & 0x3 != 0
-    {
-        return false;
-    }
-
-    // Linux's qcom,use-3-lvl-tables caps the AArch64 aperture at 39 bits;
-    // T0SZ=25 (39-bit) and T0SZ=32 (32-bit) both use an L1->L2->L3 walk for
-    // a 4 KiB granule. This checker intentionally refuses an unfamiliar
-    // format instead of guessing at a table level.
-    let iova_bits = 64u32.saturating_sub(tcr & 0x3f);
-    if !(30..=39).contains(&iova_bits) {
-        return false;
-    }
-
-    // Linux stores the context ASID in TTBR0[63:48]. It is not part of the
-    // physical page-table address and must be removed before the CPU-side
-    // identity walk below; otherwise a live Android context can appear
-    // unmapped solely because its ASID is non-zero.
-    let table_root = ttbr0 & 0x0000_ffff_ffff_f000;
-    let read_entry = |base: u64, index: u64| -> Option<u64> {
-        let offset = index.checked_mul(8)?;
-        let address = base.checked_add(offset)?;
-        (address <= usize::MAX as u64)
-            .then(|| unsafe { read_volatile(address as usize as *const u64) })
-    };
-    let maps_one_2m_block = |address: u64| -> bool {
-        let l1_index = (address >> 30) & 0x1ff;
-        let Some(l1) = read_entry(table_root, l1_index) else {
-            return false;
-        };
-        match l1 & SMMU_DESC_TYPE_MASK {
-            // An existing Android/IOMMU mapping may use a 1 GiB identity
-            // block for a large DMA aperture. It is just as valid for the
-            // preservation check as the finer-grained L2/L3 forms.
-            SMMU_DESC_BLOCK => {
-                l1 & SMMU_DESC_ADDRESS_MASK == address & !((1u64 << 30) - 1)
-                    && l1 & SMMU_DESC_VALID != 0
-            }
-            SMMU_DESC_TABLE => {
-                let l2_base = l1 & SMMU_DESC_ADDRESS_MASK;
-                let l2_index = (address >> 21) & 0x1ff;
-                let Some(l2) = read_entry(l2_base, l2_index) else {
-                    return false;
-                };
-                let block_base = address & !((1u64 << 21) - 1);
-                match l2 & SMMU_DESC_TYPE_MASK {
-                    SMMU_DESC_BLOCK => {
-                        l2 & SMMU_DESC_ADDRESS_MASK == block_base && l2 & SMMU_DESC_VALID != 0
-                    }
-                    SMMU_DESC_TABLE => {
-                        let l3_base = l2 & SMMU_DESC_ADDRESS_MASK;
-                        let l3_index = (address >> 12) & 0x1ff;
-                        let Some(l3) = read_entry(l3_base, l3_index) else {
-                            return false;
-                        };
-                        l3 & SMMU_DESC_TYPE_MASK == SMMU_DESC_TABLE
-                            && l3 & SMMU_DESC_ADDRESS_MASK == address & !0xfff
-                            && l3 & SMMU_DESC_VALID != 0
-                    }
-                    _ => false,
-                }
-            }
-            _ => false,
-        }
-    };
-
-    let mut address = iova_start & !((1u64 << 21) - 1);
-    while address < iova_end {
-        if !maps_one_2m_block(address) {
-            return false;
-        }
-        address = address.saturating_add(1u64 << 21);
-    }
-    true
-}
-
-unsafe fn smmu_tlb_sync() {
-    unsafe {
-        write_volatile(smmu_reg(SMMU_TLB_ALL_H), 0);
-        write_volatile(smmu_reg(SMMU_TLB_SYNC), 0);
-        for _ in 0..100_000u32 {
-            if read_volatile(smmu_reg(SMMU_TLB_STATUS)) & SMMU_TLB_STATUS_ACTIVE == 0 {
-                break;
-            }
-            core::arch::asm!("nop", options(nomem, nostack, preserves_flags));
-        }
-    }
-}
-
-/// Consume an Apps-SMMU context fault using the same ordering as Linux's
-/// arm_smmu_context_fault(): sample FSR/FAR/FSYNR0, retain the evidence, then
-/// clear FSR and terminate a stalled transaction. This is polled because the
-/// The normal entry is now the Apps-SMMU global/context IRQ path; polling
-/// remains as a recovery path for the short interval before the GIC owner is
-/// installed or when firmware leaves a latched fault asserted.
-unsafe fn service_smmu_fault() {
-    let global_fsr = unsafe { read_volatile(smmu_reg(SMMU_GR0_FSR)) };
-    if global_fsr != 0 && global_fsr != u32::MAX && global_fsr & SMMU_GLOBAL_FSR_FAULT != 0 {
-        let global_fsynr0 = unsafe { read_volatile(smmu_reg(SMMU_GR0_FSYNR0)) };
-        trace_event(TRACE_SMMU_GLOBAL_FAULT, global_fsr, global_fsynr0, 0, 0, 0);
-        log_hex("usb: Apps SMMU global fault FSR=", global_fsr as u64);
-        unsafe { write_volatile(smmu_reg(SMMU_GR0_FSR), global_fsr) };
-        unsafe { core::arch::asm!("dsb sy", options(nostack)) };
-    }
-    let page = unsafe { SMMU_CONTEXT_PAGE };
-    let page_size = unsafe { SMMU_CONTEXT_PAGE_SIZE };
-    if page == usize::MAX || page_size == 0 {
-        return;
-    }
-    let fsr = unsafe { smmu_page_read(page_size, page, SMMU_CB_FSR) };
-    if fsr == 0 || fsr == u32::MAX || fsr & SMMU_FSR_FAULT == 0 {
-        return;
-    }
-    let far = unsafe { smmu_page_read64(page_size, page, SMMU_CB_FAR) };
-    let fsynr0 = unsafe { smmu_page_read(page_size, page, SMMU_CB_FSYNR0) };
-    trace_event(
-        TRACE_SMMU_FAULT,
-        fsr,
-        fsynr0,
-        far as u32,
-        (far >> 32) as u32,
-        page as u32,
-    );
-    log_hex("usb: Apps SMMU context fault FSR=", fsr as u64);
-    log_hex("usb: Apps SMMU FAR=", far);
-    unsafe { smmu_page_write(page_size, page, SMMU_CB_FSR, fsr) };
-    unsafe { core::arch::asm!("dsb sy", options(nostack)) };
-    if fsr & SMMU_FSR_SS != 0 {
-        unsafe { smmu_page_write(page_size, page, SMMU_CB_RESUME, SMMU_RESUME_TERMINATE) };
-    }
-}
-
-unsafe fn install_smmu_identity_table(pool: super::platform::bramble::DmaPoolResource) -> bool {
-    let pool_base = pool.iova_base as u64;
-    let Some(pool_end) = pool_base.checked_add(pool.size as u64) else {
-        return false;
-    };
-    // This table format intentionally maps only complete 2 MiB blocks. The
-    // Android Lito pool is aligned to that granularity; refusing an unusual
-    // DT pool is safer than exposing bytes outside the declared DMA window.
-    if pool_base & 0x1f_ffff != 0
-        || pool_end & 0x1f_ffff != 0
-        || pool_end > (1u64 << 32)
-        || pool_base >= pool_end
-    {
-        return false;
-    }
-
-    unsafe {
-        for index in 0..512usize {
-            write_volatile(addr_of_mut!(SMMU_L1.0[index]), 0);
-        }
-        for table in 0..4usize {
-            for index in 0..512usize {
-                write_volatile(addr_of_mut!(SMMU_L2[table].0[index]), 0);
-            }
-        }
-
-        for l1_index in 0..4usize {
-            let l1_base = (l1_index as u64) << 30;
-            let l1_end = l1_base + (1u64 << 30);
-            if pool_base >= l1_end || pool_end <= l1_base {
-                continue;
-            }
-            // The L1 table descriptor points at the corresponding 4 KiB
-            // level-2 table. Each L2 entry describes a 2 MiB block.
-            let table_address = addr_of!(SMMU_L2[l1_index]) as usize as u64;
-            write_volatile(
-                addr_of_mut!(SMMU_L1.0[l1_index]),
-                (table_address & !0xfff) | SMMU_DESC_TABLE,
-            );
-            for l2_index in 0..512usize {
-                let physical = l1_base + (l2_index as u64) * (1u64 << 21);
-                if physical >= pool_base && physical + (1u64 << 21) <= pool_end {
-                    let descriptor = physical
-                        | SMMU_DESC_VALID
-                        | SMMU_DESC_AF
-                        | SMMU_DESC_SH_INNER
-                        | SMMU_DESC_ATTR_NORMAL
-                        | SMMU_DESC_XN;
-                    write_volatile(addr_of_mut!(SMMU_L2[l1_index].0[l2_index]), descriptor);
-                }
-            }
-        }
-        cache_clean(
-            addr_of!(SMMU_L1) as usize,
-            core::mem::size_of::<SmmuTable>(),
-        );
-        cache_clean(
-            addr_of!(SMMU_L2) as usize,
-            core::mem::size_of::<[SmmuTable; 4]>(),
-        );
-    }
-    true
-}
-
-/// Install an AArch64 stage-1 identity mapping for DWC3's stream ID.
-///
-/// Bramble's vendor DT assigns DWC3 stream ID 0xe0 to the Apps SMMU and puts
-/// USB buffers in the 0x90000000..0xf0000000 IOVA pool. Qualcomm's SMMU-500
-/// firmware can reject a direct BYPASS write by turning it into FAULT, so a
-/// real context-bank map is required here. We preserve the existing SMR and
-/// route it to a context bank configured as S1 translation + S2 bypass.
-pub fn configure_dwc3_smmu() -> bool {
-    unsafe {
-        // A failed reconfiguration must not leave the fault poller pointing
-        // at a context bank from an earlier controller lifetime.
-        SMMU_CONTEXT_PAGE = usize::MAX;
-        SMMU_CONTEXT_PAGE_SIZE = 0;
-        let pool = super::platform::bramble::usb_resources().dma_pool;
-        let dma_start = addr_of!(__usb_dma_start) as usize;
-        let dma_end = addr_of!(__usb_dma_end) as usize;
-        trace_event(
-            TRACE_SMMU_BEGIN,
-            apps_smmu_base() as u32,
-            pool.stream_id,
-            dma_start as u32,
-            dma_end as u32,
-            pool.iova_base as u32,
-        );
-        let Some(pool_end) = pool.iova_base.checked_add(pool.size) else {
-            log_puts("usb: invalid DT DMA pool\n");
-            return false;
-        };
-        if dma_start < pool.iova_base || dma_end > pool_end || dma_start >= dma_end {
-            log_puts("usb: DMA section is outside the DT IOVA pool\n");
-            return false;
-        }
-        let id0 = read_volatile(smmu_reg(SMMU_ID0));
-        let id1 = read_volatile(smmu_reg(SMMU_ID1));
-        if id0 == 0 || id0 == u32::MAX || id1 == 0 || id1 == u32::MAX {
-            log_puts("usb: Apps SMMU identification unavailable\n");
-            return false;
-        }
-
-        let num_smrs = ((id0 & SMMU_ID0_NUMSMRG_MASK) as usize).min(128);
-        let page_size = if id1 & SMMU_ID1_PAGESIZE != 0 {
-            0x10000
-        } else {
-            0x1000
-        };
-        if page_size != 0x1000 {
-            // The table below is intentionally 4 KiB-granule LPAE. Do not
-            // enable a mismatched table on a future 64 KiB-only SMMU.
-            log_puts("usb: Apps SMMU requires unsupported 64K tables\n");
-            return false;
-        }
-
-        let num_pages =
-            1usize << (((id1 >> SMMU_ID1_NUMPAGENDXB_SHIFT) & SMMU_ID1_NUMPAGENDXB_MASK) + 1);
-        let num_s2_context_banks =
-            ((id1 >> SMMU_ID1_NUMS2CB_SHIFT) & SMMU_ID1_NUMS2CB_MASK) as usize;
-        let num_context_banks = (id1 & SMMU_ID1_NUMCB_MASK) as usize;
-        if num_pages == 0 || num_context_banks == 0 {
-            log_puts("usb: Apps SMMU has no usable context banks\n");
-            return false;
-        }
-        // The GR0 window is page 0 and GR1 is page 1. Context-bank pages start
-        // after the implementation-defined number of global pages.
-        let gr1_page = 1usize;
-        let cb_base_page = num_pages;
-        log_hex("usb: Apps SMMU ID0=", id0 as u64);
-        log_hex("usb: Apps SMMU ID1=", id1 as u64);
-        log_hex("usb: Apps SMMU pages=", num_pages as u64);
-
-        let mut matched = None;
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID == 0 {
-                continue;
-            }
-            let id = smr & 0xffff;
-            let mask = (smr >> SMMU_SMR_MASK_SHIFT) & 0x7fff;
-            if ((pool.stream_id ^ id) & !mask) == 0 {
-                matched = Some((index, read_volatile(smmu_reg(SMMU_S2CR_BASE + index * 4))));
-                break;
-            }
-        }
-        let Some((smr_index, old_s2cr)) = matched else {
-            log_puts("usb: DWC3 stream 0xe0 has no SMMU match\n");
-            return false;
-        };
-
-        let old_type = old_s2cr & SMMU_S2CR_TYPE_MASK;
-        let old_cb = (old_s2cr & SMMU_S2CR_CBNDX_MASK) as usize;
-        // Fastboot commonly keeps the DWC3 stream in S2CR.BYPASS while its
-        // USB buffers are identity-addressed. This is already a valid
-        // physical=IOVA mapping for the linker DMA pool. Replacing it with a
-        // freshly selected context bank during `fastboot boot` changes an
-        // ownership boundary that Linux would preserve until its IOMMU
-        // domain is fully attached, and can make the first EP0 TRB fault.
-        if old_type == SMMU_S2CR_TYPE_BYPASS {
-            trace_event(
-                TRACE_SMMU_PRESERVED,
-                smr_index as u32,
-                old_cb as u32,
-                old_s2cr,
-                dma_start as u32,
-                dma_end as u32,
-            );
-            log_puts("usb: preserving Fastboot Apps SMMU bypass\n");
-            return true;
-        }
-        let cbndx = if old_type == SMMU_S2CR_TYPE_TRANS {
-            if old_cb >= num_context_banks || old_cb < num_s2_context_banks {
-                log_puts("usb: DWC3 SMMU context bank is out of range\n");
-                return false;
-            }
-            old_cb
-        } else {
-            // This is the same reserved-last-context-bank strategy used by
-            // Linux's qcom_smmu bypass-quirk path for firmware that refuses
-            // BYPASS S2CR values.
-            num_context_banks - 1
-        };
-        log_hex("usb: DWC3 SMMU SMR=", smr_index as u64);
-        log_hex("usb: DWC3 SMMU CB=", cbndx as u64);
-        // CBAR.IRPTNDX is an index into the dense context-bank IRQ list in
-        // the provider DT, not a GIC SPI number.  Linux's irq-domain setup
-        // programs the same association before enabling context faults.
-        let context_irq_count = super::platform::bramble::usb_resources().smmu_context_irq_count;
-        let irptndx = if context_irq_count != 0 {
-            cbndx % context_irq_count
-        } else {
-            cbndx
-        };
-        SMMU_CONTEXT_PAGE = cb_base_page + cbndx;
-        SMMU_CONTEXT_PAGE_SIZE = page_size;
-
-        // Linux's SMMU driver owns this context, not the DWC3 glue. A
-        // Fastboot handoff can therefore arrive with a valid stage-1 map
-        // already installed for the same stream. Replacing that context
-        // underneath firmware is unsafe: the controller may still have an
-        // outstanding transaction using the old page tables, and changing
-        // TTBR0 can turn a benign handoff into a stream fault. Preserve a
-        // live translation context and use the declared pool check above as
-        // the ownership boundary for Fullerene's DMA objects.
-        if old_type == SMMU_S2CR_TYPE_TRANS {
-            let cb_page = cb_base_page + cbndx;
-            let sctlr = smmu_page_read(page_size, cb_page, SMMU_CB_SCTLR);
-            let ttbr0 = smmu_page_read64(page_size, cb_page, SMMU_CB_TTBR0);
-            let tcr = smmu_page_read(page_size, cb_page, SMMU_CB_TCR);
-            if sctlr & SMMU_SCTLR_M != 0
-                && smmu_context_maps_identity(ttbr0, tcr, dma_start, dma_end)
-            {
-                trace_event(
-                    TRACE_SMMU_PRESERVED,
-                    smr_index as u32,
-                    cbndx as u32,
-                    sctlr,
-                    ttbr0 as u32,
-                    (ttbr0 >> 32) as u32,
-                );
-                log_puts("usb: preserving active Apps SMMU translation\n");
-                return true;
-            }
-            log_puts("usb: active Apps SMMU map does not cover Fullerene DMA\n");
-        }
-
-        if !install_smmu_identity_table(pool) {
-            log_puts("usb: DT DMA pool is not 2 MiB aligned/32-bit addressable\n");
-            return false;
-        }
-
-        // Stop the bank before changing its format and page-table pointer.
-        smmu_page_write(page_size, cb_base_page + cbndx, SMMU_CB_SCTLR, 0);
-        smmu_page_write(
-            page_size,
-            gr1_page,
-            SMMU_GR1_CBA2R_BASE + cbndx * 4,
-            SMMU_CBA2R_VA64,
-        );
-        smmu_page_write(
-            page_size,
-            gr1_page,
-            SMMU_GR1_CBAR_BASE + cbndx * 4,
-            SMMU_CBAR_S1_TRANS_S2_BYPASS
-                | SMMU_CBAR_S1_MEMATTR_WB
-                | SMMU_CBAR_S1_BPSHCFG_NSH
-                | (irptndx as u32 & SMMU_CBAR_IRPTNDX_MASK),
-        );
-
-        // SCR0 interrupt enables are deliberately NOT written here. The
-        // global/CFG fault lines are polled (service_smmu_fault) and the
-        // secure-side SCR0 bits can reject non-secure writes on Qualcomm
-        // firmware, which aborts the whole handoff before the pull-up.
-
-        let cb_page = cb_base_page + cbndx;
-        // 4 KiB granule, 32-bit IOVA, inner-shareable WBWA walks, and a
-        // 40-bit output address size. TCR2 selects the AArch64 format.
-        smmu_page_write(
-            page_size,
-            cb_page,
-            SMMU_CB_TCR2,
-            SMMU_TCR2_SEP_UPSTREAM | SMMU_TCR2_AS | SMMU_TCR2_PASIZE_40BIT,
-        );
-        let t0sz = if super::platform::bramble::usb_resources().smmu_use_3_level_tables {
-            SMMU_TCR_T0SZ_39BIT
-        } else {
-            SMMU_TCR_T0SZ_32BIT
-        };
-        smmu_page_write(
-            page_size,
-            cb_page,
-            SMMU_CB_TCR,
-            SMMU_TCR_EPD1 | SMMU_TCR_SH0_INNER | SMMU_TCR_ORGN0_WBWA | SMMU_TCR_IRGN0_WBWA | t0sz,
-        );
-        smmu_page_write64(
-            page_size,
-            cb_page,
-            SMMU_CB_TTBR0,
-            addr_of!(SMMU_L1) as usize as u64,
-        );
-        smmu_page_write64(page_size, cb_page, SMMU_CB_TTBR1, 0);
-        smmu_page_write(page_size, cb_page, SMMU_CB_CONTEXTIDR, 0);
-        smmu_page_write(page_size, cb_page, SMMU_CB_MAIR0, 0xff);
-        smmu_page_write(page_size, cb_page, SMMU_CB_MAIR1, 0);
-        smmu_page_write(
-            page_size,
-            cb_page,
-            SMMU_CB_SCTLR,
-            SMMU_SCTLR_S1_ASIDPNE
-                | SMMU_SCTLR_CFIE
-                | SMMU_SCTLR_CFRE
-                | SMMU_SCTLR_AFE
-                | SMMU_SCTLR_TRE
-                | SMMU_SCTLR_M,
-        );
-
-        // S2CR type TRANS is zero; preserve privilege and EXID bits from the
-        // firmware entry while replacing only the context-bank selector.
-        let new_s2cr = (old_s2cr & !SMMU_S2CR_CBNDX_MASK) & !SMMU_S2CR_TYPE_MASK
-            | ((cbndx as u32) & SMMU_S2CR_CBNDX_MASK)
-            | SMMU_S2CR_TYPE_TRANS;
-        let s2cr_address = SMMU_S2CR_BASE + smr_index * 4;
-        write_volatile(smmu_reg(s2cr_address), new_s2cr);
-        core::arch::asm!("dsb sy", options(nostack));
-        let readback = read_volatile(smmu_reg(s2cr_address));
-        if readback & SMMU_S2CR_TYPE_MASK != SMMU_S2CR_TYPE_TRANS
-            || (readback & SMMU_S2CR_CBNDX_MASK) as usize != cbndx
-        {
-            log_puts("usb: DWC3 SMMU S2CR translation rejected\n");
-            return false;
-        }
-        smmu_tlb_sync();
-        trace_event(
-            TRACE_SMMU_READY,
-            smr_index as u32,
-            cbndx as u32,
-            id0,
-            id1,
-            0,
-        );
-        true
-    }
-}
-
-#[inline]
-unsafe fn read_qscratch(offset: usize) -> u32 {
-    unsafe { read_volatile(qscratch_reg(offset)) }
-}
-
-#[inline]
-unsafe fn write_qscratch(offset: usize, value: u32) {
-    unsafe { write_volatile(qscratch_reg(offset), value) };
-    let _ = unsafe { read_volatile(qscratch_reg(offset)) };
-}
-
-#[inline]
-unsafe fn hsphy_update(offset: usize, mask: u32, value: u32) {
-    let current = unsafe { read_volatile(hsphy_reg(offset)) };
-    unsafe { write_volatile(hsphy_reg(offset), (current & !mask) | (value & mask)) };
-    let _ = unsafe { read_volatile(hsphy_reg(offset)) };
-}
-
-unsafe fn init_qmp_phy() -> bool {
-    let com_power_down = qmp_contract_offset(8, QMP_COM_POWER_DOWN_CTRL);
-    let pcs_power_down = qmp_contract_offset(3, QMP_PCS_POWER_DOWN_CONTROL);
-    let reset_override = qmp_contract_offset(10, QMP_COM_RESET_OVRD_CTRL);
-    let typec = qmp_contract_offset(12, QMP_COM_TYPEC_CTRL);
-    let phy_mode = qmp_contract_offset(11, QMP_COM_PHY_MODE_CTRL);
-    let com_sw_reset = qmp_contract_offset(9, QMP_COM_SW_RESET);
-    let pcs_sw_reset = qmp_contract_offset(4, QMP_PCS_SW_RESET);
-    let pcs_start = qmp_contract_offset(5, QMP_PCS_START_CONTROL);
-    let pcs_status = qmp_contract_offset(0, QMP_PCS_STATUS1);
-    unsafe {
-        // Match msm_ssphy_qmp_init(): power the common and PCS blocks before
-        // selecting the Type-C lane and USB+DP combo mode. The lane value is
-        // 2 for lane A and 3 for lane B, as used by the Android QMP driver.
-        write_volatile(qmp_reg(com_power_down), 0x01);
-        write_volatile(qmp_reg(pcs_power_down), 0x01);
-        let lane = if TYPEC_LANE_B { 0x03 } else { 0x02 };
-        write_volatile(qmp_reg(reset_override), 0x0f);
-        write_volatile(qmp_reg(typec), lane);
-        let _ = read_volatile(qmp_reg(typec));
-        write_volatile(qmp_reg(phy_mode), 0x03);
-        let _ = read_volatile(qmp_reg(phy_mode));
-        write_volatile(qmp_reg(reset_override), 0x00);
-
-        for index in 0..146 {
-            let (offset, value) = ACTIVE_QMP_INIT[index];
-            write_volatile(qmp_reg(offset), value);
-            let delay_us = ACTIVE_QMP_INIT_DELAY_US[index];
-            if delay_us != 0 {
-                super::timer::delay_us(delay_us as u64);
-            }
-        }
-
-        write_volatile(qmp_reg(com_sw_reset), 0x00);
-        write_volatile(qmp_reg(pcs_sw_reset), 0x00);
-        write_volatile(qmp_reg(pcs_start), 0x03);
-        let _ = read_volatile(qmp_reg(pcs_status));
-        for _ in 0..1_000_000 {
-            if read_volatile(qmp_reg(pcs_status)) & QMP_PHYSTATUS == 0 {
-                return true;
-            }
-            core::arch::asm!("nop", options(nomem, nostack, preserves_flags));
-        }
-    }
-    log_puts("usb: QMP PHY initialization timeout\n");
-    false
-}
-
-/// Clear the QMP LFPS receiver-detect interrupt using the required 1 -> 0
-/// sequence from msm-ssusb-qmp. A readback between the writes is not needed
-/// by the PHY, but the compiler/MMIO ordering barrier is: the second write
-/// must not be observed before the clear is asserted.
-unsafe fn qmp_clear_lfps_rxterm_irq() {
-    let clear = qmp_contract_offset(2, QMP_PCS_LFPS_RXTERM_IRQ_CLEAR);
-    unsafe {
-        write_volatile(qmp_reg(clear), QMP_LFPS_IRQ_CLEAR);
-        core::sync::atomic::fence(core::sync::atomic::Ordering::SeqCst);
-        write_volatile(qmp_reg(clear), 0);
-        let _ = read_volatile(qmp_reg(clear));
-    }
-}
-
-/// Match msm_ssusb_qmp_enable_autonomous()/disable_autonomous_mode() for the
-/// Lito USB+DP combo PHY. The device-side SuperSpeed mode enables both
-/// receiver-detect and LFPS detection; the receiver-detect event-select bit
-/// stays clear in that mode. Autonomous mode also turns on the PCS I/O clamp
-/// (the register is active-high for disabling the clamp, hence clear it when
-/// enabling autonomous operation).
-unsafe fn qmp_set_autonomous_mode(enable: bool) {
-    let autonomous = qmp_contract_offset(1, QMP_PCS_AUTONOMOUS_MODE_CTRL);
-    let clamp_offset = qmp_contract_offset(14, QMP_PCS_CLAMP_ENABLE);
-    unsafe {
-        if enable {
-            qmp_clear_lfps_rxterm_irq();
-            let mut value = read_volatile(qmp_reg(autonomous));
-            value &= !(QMP_ARCVR_DTCT_EN | QMP_ALFPS_DTCT_EN | QMP_ARCVR_DTCT_EVENT_SEL);
-            value |= QMP_ARCVR_DTCT_EN | QMP_ALFPS_DTCT_EN;
-            write_volatile(qmp_reg(autonomous), value);
-            // Android's combo-PHY path calls clamp_enable(true), which
-            // writes !true to this active-high clamp control.
-            let mut clamp = read_volatile(qmp_reg(clamp_offset));
-            clamp &= !QMP_CLAMP_EN;
-            write_volatile(qmp_reg(clamp_offset), clamp);
-            let _ = read_volatile(qmp_reg(autonomous));
-        } else {
-            // Resume first releases the clamp, then disables both autonomous
-            // detectors, and finally clears any receiver-detect edge left by
-            // the suspended PHY.
-            let mut clamp = read_volatile(qmp_reg(clamp_offset));
-            clamp |= QMP_CLAMP_EN;
-            write_volatile(qmp_reg(clamp_offset), clamp);
-            let mut value = read_volatile(qmp_reg(autonomous));
-            value &= !(QMP_ARCVR_DTCT_EN | QMP_ALFPS_DTCT_EN | QMP_ARCVR_DTCT_EVENT_SEL);
-            write_volatile(qmp_reg(autonomous), value);
-            qmp_clear_lfps_rxterm_irq();
-        }
-    }
-}
-
-/// Apply the small, non-calibration portion of the SM7250 USB2 PHY setup.
-///
-/// The full Linux driver also obtains regulators and a 19.2 MHz reference
-/// clock from the board description. Those are already left on by the
-/// Pixel boot chain; the register sequence below is the actual PHY setup
-/// from the `qcom,usb-hsphy-snps-femto` driver and the Bramble override
-/// sequence in its device tree.
-unsafe fn init_hsphy() {
-    unsafe {
-        hsphy_update(
-            HSPHY_CFG0,
-            HSPHY_CFG0_CMN_CTRL_OVERRIDE_EN,
-            HSPHY_CFG0_CMN_CTRL_OVERRIDE_EN,
-        );
-        hsphy_update(HSPHY_UTMI_CTRL5, HSPHY_UTMI_POR, HSPHY_UTMI_POR);
-        hsphy_update(HSPHY_COMMON0, HSPHY_COMMON0_FSEL_MASK, 0);
-        hsphy_update(
-            HSPHY_COMMON1,
-            HSPHY_COMMON1_PLLBTUNE,
-            HSPHY_COMMON1_PLLBTUNE,
-        );
-        hsphy_update(HSPHY_REFCLK_CTRL, 0x3, 0x2);
-        hsphy_update(
-            HSPHY_COMMON1,
-            HSPHY_COMMON1_VBUSVLDEXTSEL0,
-            HSPHY_COMMON1_VBUSVLDEXTSEL0,
-        );
-        hsphy_update(
-            HSPHY_CTRL1,
-            HSPHY_CTRL1_VBUSVLDEXT0,
-            HSPHY_CTRL1_VBUSVLDEXT0,
-        );
-
-        // qcom,param-override-seq is encoded as (value, register offset).
-        for index in 0..3 {
-            let (offset, value) = ACTIVE_HSPHY_PARAM_OVERRIDE[index];
-            write_volatile(hsphy_reg(offset), value);
-        }
-
-        // Bramble does not declare an external-calibration resistor, so the
-        // upstream driver enables the internal RTUNE path.
-        hsphy_update(HSPHY_RTUNE_SEL, 1, 1);
-        hsphy_update(
-            HSPHY_COMMON2,
-            HSPHY_COMMON2_VREGBYPASS,
-            HSPHY_COMMON2_VREGBYPASS,
-        );
-        // The SNPS Femto driver uses the ATE/test toggle sequence to commit
-        // the PHY's analog override values before releasing POR.
-        hsphy_update(HSPHY_UTMI_CTRL5, HSPHY_UTMI_ATE_RESET, HSPHY_UTMI_ATE_RESET);
-        hsphy_update(
-            HSPHY_TEST1,
-            HSPHY_TEST1_TESTDATAOUTSEL,
-            HSPHY_TEST1_TESTDATAOUTSEL,
-        );
-        hsphy_update(HSPHY_TEST1, HSPHY_TEST1_TOGGLE_2WR, HSPHY_TEST1_TOGGLE_2WR);
-        hsphy_update(HSPHY_COMMON0, HSPHY_COMMON0_VATESTENB_MASK, 0);
-        hsphy_update(HSPHY_TEST0, HSPHY_TEST0_DATA_MASK, 0);
-        hsphy_update(
-            HSPHY_CTRL2,
-            HSPHY_CTRL2_SUSPEND_N_SEL | HSPHY_CTRL2_SUSPEND_N,
-            HSPHY_CTRL2_SUSPEND_N_SEL | HSPHY_CTRL2_SUSPEND_N,
-        );
-        hsphy_update(HSPHY_UTMI_CTRL0, HSPHY_UTMI_SLEEPM, HSPHY_UTMI_SLEEPM);
-        hsphy_update(HSPHY_UTMI_CTRL5, HSPHY_UTMI_POR, 0);
-        hsphy_update(HSPHY_CTRL2, HSPHY_CTRL2_SUSPEND_N_SEL, 0);
-        hsphy_update(HSPHY_CFG0, HSPHY_CFG0_CMN_CTRL_OVERRIDE_EN, 0);
-    }
-}
-
-unsafe fn select_utmi_pipe_clock() {
-    // This is the Qualcomm glue sequence used when DWC3 operates without a
-    // SuperSpeed PHY. It prevents the absent QMP PIPE clock from holding the
-    // core in reset while the USB2 UTMI clock is already running.
-    trace_event(TRACE_UTMI_CLOCK, 0, 0, 0, 0, 0);
-    unsafe {
-        qscratch_set(QSCRATCH_GENERAL_CFG, PIPE_UTMI_CLK_DIS);
-        // dwc3_qcom_select_utmi_clk() uses usleep_range(100, 1000) between
-        // each clock-source transition.  A fixed architectural delay keeps
-        // the lower bound independent of the boot CPU frequency; a NOP loop
-        // could be shorter than 100 us on a fast handset.
-        super::timer::delay_us(100);
-        qscratch_set(QSCRATCH_GENERAL_CFG, PIPE_UTMI_CLK_SEL | PIPE3_PHYSTATUS_SW);
-        super::timer::delay_us(100);
-        let value = read_qscratch(QSCRATCH_GENERAL_CFG) & !PIPE_UTMI_CLK_DIS;
-        write_qscratch(QSCRATCH_GENERAL_CFG, value);
-    }
-    trace_event(TRACE_UTMI_CLOCK, 1, 0, 0, 0, 0);
-}
-
-/// Apply the DWC3 post-reset reference-clock calibration from
-/// dwc3_msm_update_ref_clk(). The GCC source clock is managed separately by
-/// the Bramble platform layer; this only programs the controller's timing
-/// registers after a core reset.
-unsafe fn update_dwc3_ref_clock() {
-    unsafe {
-        let guctl = read(GUCTL);
-        write(
-            GUCTL,
-            (guctl & !GUCTL_REFCLKPER_MASK) | GUCTL_REFCLKPER_19_2MHZ,
-        );
-        if read(GSNPSID) >= DWC3_REVISION_250A {
-            let gfladj = read(GFLADJ);
-            write(
-                GFLADJ,
-                (gfladj
-                    & !(GFLADJ_REFCLK_FLADJ_MASK
-                        | GFLADJ_REFCLK_LPM_SEL
-                        | GFLADJ_REFCLK_240MHZ_DECR
-                        | GFLADJ_REFCLK_240MHZDECR_PLS1))
-                    | GFLADJ_REFCLK_LPM_SEL
-                    | GFLADJ_REFCLK_240MHZ_DECR
-                    | GFLADJ_REFCLK_240MHZDECR_PLS1
-                    | GFLADJ_REFCLK_FLADJ_19_2MHZ,
-            );
-        }
-    }
-}
-
-/// Reset the DWC3 core after taking ownership from the bootloader.
-///
-/// The Qualcomm glue invokes this as part of the DWC3 post-reset path. A
-/// `fastboot boot` handoff skips that driver, so leaving the controller in its
-/// bootloader device/host state can make endpoint commands retire without
-/// ever allowing the peripheral pull-up to become visible.
-unsafe fn device_soft_reset() -> bool {
-    unsafe {
-        trace_event(TRACE_DWC3_RESET_BEGIN, 0, 0, 0, 0, 0);
-        trace_event(TRACE_DEVICE_RESET, 0, 0, 0, 0, 0);
-        let initial_dctl = read(DCTL);
-        // Match Linux's reconnect path: clear stale endpoint/device state
-        // without touching the already-running Qualcomm PHY and clock
-        // branches. RUN_STOP must be cleared in the same write; preserving
-        // Fastboot's RUN_STOP bit can leave the device half-running while
-        // CSFTRST is asserted.
-        let mut dctl = initial_dctl;
-        dctl |= DCTL_CSFTRST;
-        dctl &= !DCTL_RUN_STOP;
-        write_dctl_safe(dctl);
-        let snpsid = read(GSNPSID);
-        let ip = snpsid >> 16;
-        // DWC_usb31 1.90a+ and DWC_usb32 synchronize CSFTRST through all
-        // clocks and need the slower 20-ms polling cadence used by Linux.
-        // Bramble's 0x5533 controller follows the ordinary 1-us path.
-        let version = if ip == DWC31_IP || ip == DWC32_IP {
-            read(VER_NUMBER)
-        } else {
-            0
-        };
-        let slow_reset = ip == DWC32_IP || (ip == DWC31_IP && version >= DWC31_REVISION_190A);
-        let retries = if slow_reset { 10 } else { 1_000 };
-        let mut device_reset_complete = false;
-        for _ in 0..retries {
-            if read(DCTL) & DCTL_CSFTRST == 0 {
-                device_reset_complete = true;
-                break;
-            }
-            if slow_reset {
-                super::timer::delay_ms(20);
-            } else {
-                super::timer::delay_us(1);
-            }
-        }
-        if !device_reset_complete {
-            log_puts("usb: DWC3 device reset timeout\n");
-            return false;
-        }
-
-        // Upstream Linux waits an additional 50 ms only for DWC_usb31 1.80a
-        // and earlier before accessing its PHY domain. DWC3/1.90a+ do not
-        // require that legacy synchronization delay.
-        if ip == DWC31_IP && version <= DWC31_REVISION_180A {
-            super::timer::delay_ms(50);
-        }
-        true
-    }
-}
-
-/// Mirror Linux's dwc3_gadget_dctl_write_safe(). DCTL's link-state request
-/// field is a command, not persistent configuration; carrying a Fastboot
-/// request into CSFTRST or Run/Stop can make the next device transition race
-/// the controller's link state machine.
-#[inline]
-unsafe fn write_dctl_safe(value: u32) {
-    unsafe { write(DCTL, value & !DCTL_TRGTULST_MASK) };
-}
-
-/// Reset the DWC3 core and both PHY-facing domains for a cold platform start.
-///
-/// This is intentionally separate from `device_soft_reset`: a Fastboot
-/// handoff must not reset the PHYs that own the Type-C session.
-unsafe fn core_soft_reset(super_speed: bool) -> bool {
-    unsafe {
-        if !device_soft_reset() {
-            return false;
-        }
-
-        // B3a: the CSFTRST handshake completed (DCTL.CSFTRST cleared) but
-        // the GCTL.CORESOFTRESET / PHY_PHYSOFTRST section has not started.
-        #[cfg(fullerene_aarch64_usb_ep0_signal_probe)]
-        init_beacon();
-
-        let mut gctl = read(GCTL);
-        gctl |= GCTL_CORESOFTRESET;
-        write(GCTL, gctl);
-
-        let mut usb2 = read(GUSB2PHYCFG0);
-        usb2 |= GUSB2PHYCFG_PHYSOFTRST;
-        write(GUSB2PHYCFG0, usb2);
-        if super_speed {
-            let mut usb3 = read(GUSB3PIPECTL0);
-            usb3 |= GUSB3PIPECTL_PHYSOFTRST;
-            write(GUSB3PIPECTL0, usb3);
-        }
-
-        // The upstream DWC3 core reset uses a 100 ms delay after releasing
-        // both PHY resets. The architectural counter is firmware-provided
-        // before this early probe, so use it instead of a CPU-dependent loop.
-        super::timer::delay_ms(100);
-
-        usb2 = read(GUSB2PHYCFG0) & !GUSB2PHYCFG_PHYSOFTRST;
-        write(GUSB2PHYCFG0, usb2);
-        if super_speed {
-            let mut usb3 = read(GUSB3PIPECTL0);
-            usb3 &= !GUSB3PIPECTL_PHYSOFTRST;
-            write(GUSB3PIPECTL0, usb3);
-        }
-        super::timer::delay_ms(1);
-
-        gctl = read(GCTL) & !GCTL_CORESOFTRESET;
-        write(GCTL, gctl);
-        true
-    }
-}
-
-/// Stop a controller that was left running by Fastboot before reusing its
-/// device-mode endpoint state. A DWC3 gadget must be halted before
-/// DEPSTARTCFG/SETEPCONFIG are issued; a handoff cannot assume that the
-/// bootloader performed the normal gadget-stop sequence.
-unsafe fn stop_running_device() -> bool {
-    unsafe { run_stop_device(false) }
-}
-
-const DWC3_RUN_STOP_POLL_MS: u64 = 1;
-const DWC3_RUN_STOP_TIMEOUT_MS: u64 = 2_000;
 // STARTTRANSFER must DMA-fetch the TRB before the command can retire, and on
 // this platform that first fetch is far slower than the other endpoint
 // commands (which complete from the register path alone). The probe-era
@@ -3415,95 +1059,6 @@ const DWC3_RUN_STOP_TIMEOUT_MS: u64 = 2_000;
 // 2,000,000 reads bounds the wait at a comfortable fraction of a second
 // without ever spinning forever.
 const DWC3_EP_COMMAND_TIMEOUT: u32 = 2_000_000;
-
-/// Wait for DWC3's device controller to reach the requested halt state after
-/// a Run/Stop write. Linux polls DSTS at 1--2 ms intervals for up to 2,000
-/// iterations; a fixed NOP count is too short on a fast boot CPU and can let
-/// endpoint commands race the controller's previous Fastboot session.
-unsafe fn wait_device_state(want_halted: bool) -> bool {
-    unsafe {
-        for _ in 0..DWC3_RUN_STOP_TIMEOUT_MS {
-            // The DWC3 databook requires software to acknowledge device
-            // events while waiting for DEVCTRLHLT during a gadget stop.
-            // Linux's soft-disconnect path does this in parallel with the
-            // halt poll; a Fastboot-owned stale disconnect/reset event can
-            // otherwise keep the controller from completing Run/Stop.
-            if want_halted {
-                acknowledge_events_while_halting();
-            }
-            let dsts = read(DSTS);
-            let halted = dsts & DSTS_DEVCTRLHLT != 0;
-            if halted == want_halted {
-                if want_halted {
-                    trace_event(TRACE_DWC3_HALTED, 0, 0, 0, 0, dsts);
-                }
-                return true;
-            }
-            super::timer::delay_ms(DWC3_RUN_STOP_POLL_MS);
-        }
-        let dsts = read(DSTS);
-        trace_event(TRACE_DWC3_HALT_TIMEOUT, want_halted as u32, 0, 0, 0, dsts);
-        log_hex(
-            if want_halted {
-                "usb: DWC3 stop timeout during handoff, DSTS="
-            } else {
-                "usb: DWC3 start timeout during handoff, DSTS="
-            },
-            dsts as u64,
-        );
-        false
-    }
-}
-
-/// Acknowledge events generated while DWC3 is draining a device stop.
-///
-/// This is intentionally separate from `acknowledge_ep0_event_count()`: event
-/// buffer setup preserves the complete GEVNTCOUNT register (including EHB),
-/// while the Run/Stop halt contract consumes only the byte count and advances
-/// the software cursor just as Linux advances `ev_buf->lpos`.
-unsafe fn acknowledge_events_while_halting() {
-    unsafe {
-        let count = read(GEVNTCOUNT0) & GEVNTCOUNT_MASK;
-        if count != 0 {
-            write(GEVNTCOUNT0, count);
-            EVENT_OFFSET = (EVENT_OFFSET + count as usize) % ep0_event_size();
-            core::arch::asm!("dsb sy", options(nostack));
-        }
-    }
-}
-
-/// Apply Linux's USB2 PHY guard around a DWC3 Run/Stop transition.
-///
-/// `dwc3_gadget_run_stop()` clears SUSPHY and ENBLSLPM before writing DCTL,
-/// waits for DEVCTRLHLT, and restores the saved bits afterwards. Keeping that
-/// sequence in one helper prevents the Fastboot handoff and runtime-PM paths
-/// from diverging at exactly the transition where DWC3 is most sensitive to
-/// a stale USB2 low-power state.
-unsafe fn run_stop_device(is_on: bool) -> bool {
-    unsafe {
-        let mut usb2 = read(GUSB2PHYCFG0);
-        let saved_config = usb2 & (GUSB2PHYCFG_SUSPHY | GUSB2PHYCFG_ENBLSLPM);
-        if saved_config != 0 {
-            usb2 &= !(GUSB2PHYCFG_SUSPHY | GUSB2PHYCFG_ENBLSLPM);
-            write(GUSB2PHYCFG0, usb2);
-        }
-
-        let mut dctl = read(DCTL);
-        if is_on {
-            dctl = run_stop_value(dctl, read(GSNPSID));
-        } else {
-            dctl &= !DCTL_RUN_STOP;
-        }
-        write_dctl_safe(dctl);
-        let complete = wait_device_state(!is_on);
-
-        if saved_config != 0 {
-            let current = read(GUSB2PHYCFG0);
-            write(GUSB2PHYCFG0, current | saved_config);
-        }
-        complete
-    }
-}
 
 #[inline]
 fn gadget_speed_value(mut dcfg: u32, super_speed: bool, snpsid: u32) -> u32 {
@@ -3749,16 +1304,6 @@ unsafe fn configure_gadget_start_defaults() {
 }
 
 #[inline]
-unsafe fn read(offset: usize) -> u32 {
-    unsafe { read_volatile(reg(offset)) }
-}
-
-#[inline]
-unsafe fn write(offset: usize, value: u32) {
-    unsafe { write_volatile(reg(offset), value) }
-}
-
-#[inline]
 unsafe fn qscratch_set(offset: usize, mask: u32) {
     trace_event(TRACE_QSCRATCH_BEGIN, offset as u32, mask, 0, 0, 0);
     let value = unsafe { read_qscratch(offset) } | mask;
@@ -3766,11 +1311,6 @@ unsafe fn qscratch_set(offset: usize, mask: u32) {
     // The QCOM glue driver performs a readback to make the peripheral-mode
     // session vote visible before it starts the DWC3 core.
     let _ = unsafe { read_qscratch(offset) };
-}
-
-#[inline]
-unsafe fn dep_reg(endpoint: usize, offset: usize) -> usize {
-    DEP_BASE + endpoint * 0x10 + offset
 }
 
 #[inline]
@@ -6258,8 +3798,7 @@ pub fn init_usb2_pullup_handoff() -> bool {
 /// cumulative cost of the executed prefix: it separates the
 /// ABL-to-kernel-entry latency from the per-step controller cost.
 fn bare_pullup_stop_after() -> Option<u32> {
-    option_env!("FULLERENE_USB_BARE_PULLUP_STOP_AFTER")
-        .and_then(|value| value.parse::<u32>().ok())
+    option_env!("FULLERENE_USB_BARE_PULLUP_STOP_AFTER").and_then(|value| value.parse::<u32>().ok())
 }
 
 unsafe fn init_usb2_bare_pullup_handoff_inner(connect: bool) -> bool {
@@ -8306,43 +5845,6 @@ pub fn ep0_raw_link_nibble() -> u32 {
     }
 }
 
-/// Read-only Apps-SMMU state for the DWC3 stream. This never writes the
-/// SMMU; if the aperture is clock-gated or secure-owned the access aborts and
-/// the handset reboots before the pull-up is published, which is itself a
-/// distinct host-visible outcome.
-///   6 = SMR matched the stream and S2CR selects TRANSLATE
-///   7 = SMR matched the stream and S2CR selects BYPASS
-///   8 = SMMU readable but no SMR matches the DWC3 stream
-///   9 = SMMU identification registers are unreadable
-pub fn probe_smmu_stream_state() -> u32 {
-    unsafe {
-        let id0 = read_volatile(smmu_reg(SMMU_ID0));
-        let id1 = read_volatile(smmu_reg(SMMU_ID1));
-        if id0 == 0 || id0 == u32::MAX || id1 == 0 || id1 == u32::MAX {
-            return 9;
-        }
-        let num_smrs = ((id0 & SMMU_ID0_NUMSMRG_MASK) as usize).min(128);
-        let stream_id = super::platform::bramble::usb_resources().dma_pool.stream_id;
-        for index in 0..num_smrs {
-            let smr = read_volatile(smmu_reg(SMMU_SMR_BASE + index * 4));
-            if smr & SMMU_SMR_VALID == 0 {
-                continue;
-            }
-            let id = smr & 0xffff;
-            let mask = (smr >> SMMU_SMR_MASK_SHIFT) & 0x7fff;
-            if (stream_id ^ id) & !mask == 0 {
-                let s2cr = read_volatile(smmu_reg(SMMU_S2CR_BASE + index * 4));
-                return if s2cr & SMMU_S2CR_TYPE_MASK == SMMU_S2CR_TYPE_BYPASS {
-                    7
-                } else {
-                    6
-                };
-            }
-        }
-        8
-    }
-}
-
 /// Heartbeat control: toggle DCTL Run/Stop in one-second intervals starting
 /// immediately after the verified connect. If the host still records a full
 /// 5-second descriptor timeout against a continuously attached port, the
@@ -8468,9 +5970,7 @@ pub fn link_on_sample() {
         }
         if !LNK_MID_SEEN {
             let state = (dsts >> 18) & 0xf;
-            if state == 8 || state == 9 || state == 11 || state == 14
-                || state == 15
-            {
+            if state == 8 || state == 9 || state == 11 || state == 14 || state == 15 {
                 LNK_MID_SEEN = true;
             }
         }
